Fish catch data set

There are 159 fishes of 7 species are caught and measured. Altogether there are 7 variables. All the fishes are caught from the same lake(Laengelmavesi) near Tampere in Finland.

Type

has 7 fish classes, with 35 cases of Bream, 11 cases of Parkki, 56 cases of Perch 17 cases of Pike, 20 cases of Roach, 14 cases of Smelt and 6 cases of Whitewish.

weight

Weight of the fish (in grams)

length1

Length from the nose to the beginning of the tail (in cm)

length2

Length from the nose to the notch of the tail (in cm)

length3

Length from the nose to the end of the tail (in cm)

height

Maximal height as % of Length3

width

Maximal width as % of Length3

Usage

data(fishcatch)

Format

A data frame with 159 rows and 7 variables

Source

[http://www.amstat.org/publications/jse/jse_data_archive.htm](fishcatch)