Data structure with the projected and boundary by node and class.

Data structure with the projected and boundary by node and class.

Usage

node_data(ppf, tr, Rule = 1)

Arguments

ppf

is a PPforest object

tr

numerical value to identify a tree

Rule

split rule 1:mean of two group means, 2:weighted mean, 3: mean of max(left group) and min(right group), 4: weighted mean of max(left group) and min(right group)

Value

Data frame with projected data for each class and node id and the boundaries

Examples

#crab data set with all the observations used as training

pprf.crab <- PPforest(data = crab, std =TRUE, class = 'Type',
 size.tr = 1, m = 200, size.p = .5, PPmethod = 'LDA')
node_data(ppf = pprf.crab, tr = 1) 
#>                proj.data        Class          cut node.id LR.class   Dir
#> 100...1     -0.136705852   BlueFemale  0.007882217       1        L FALSE
#> 84...2      -0.080607983   BlueFemale  0.007882217       1        L FALSE
#> 71...3      -0.028568960   BlueFemale  0.007882217       1        L FALSE
#> 67...4      -0.116914319   BlueFemale  0.007882217       1        L FALSE
#> 92...5      -0.184399798   BlueFemale  0.007882217       1        L FALSE
#> 58...6      -0.092855685   BlueFemale  0.007882217       1        L FALSE
#> 60...7      -0.166550064   BlueFemale  0.007882217       1        L FALSE
#> 95...8      -0.273875326   BlueFemale  0.007882217       1        L FALSE
#> 66...9      -0.082019930   BlueFemale  0.007882217       1        L FALSE
#> 69...10     -0.040858575   BlueFemale  0.007882217       1        L FALSE
#> 90...11     -0.069867755   BlueFemale  0.007882217       1        L FALSE
#> 60.1...12   -0.166550064   BlueFemale  0.007882217       1        L FALSE
#> 51...13     -0.056115025   BlueFemale  0.007882217       1        L FALSE
#> 71.1...14   -0.028568960   BlueFemale  0.007882217       1        L FALSE
#> 75...15     -0.261867490   BlueFemale  0.007882217       1        L FALSE
#> 72...16     -0.154493765   BlueFemale  0.007882217       1        L FALSE
#> 68...17     -0.122454181   BlueFemale  0.007882217       1        L FALSE
#> 94...18     -0.090201539   BlueFemale  0.007882217       1        L FALSE
#> 73...19     -0.118368179   BlueFemale  0.007882217       1        L FALSE
#> 77...20     -0.145536590   BlueFemale  0.007882217       1        L FALSE
#> 99...21     -0.066184817   BlueFemale  0.007882217       1        L FALSE
#> 89...22     -0.138770683   BlueFemale  0.007882217       1        L FALSE
#> 61...23     -0.237071448   BlueFemale  0.007882217       1        L FALSE
#> 66.1...24   -0.082019930   BlueFemale  0.007882217       1        L FALSE
#> 99.1...25   -0.066184817   BlueFemale  0.007882217       1        L FALSE
#> 80...26     -0.147807580   BlueFemale  0.007882217       1        L FALSE
#> 89.1...27   -0.138770683   BlueFemale  0.007882217       1        L FALSE
#> 69.1...28   -0.040858575   BlueFemale  0.007882217       1        L FALSE
#> 89.2...29   -0.138770683   BlueFemale  0.007882217       1        L FALSE
#> 77.1...30   -0.145536590   BlueFemale  0.007882217       1        L FALSE
#> 96...31     -0.068970902   BlueFemale  0.007882217       1        L FALSE
#> 60.2...32   -0.166550064   BlueFemale  0.007882217       1        L FALSE
#> 65...33     -0.118145516   BlueFemale  0.007882217       1        L FALSE
#> 55...34     -0.104004478   BlueFemale  0.007882217       1        L FALSE
#> 61.1...35   -0.237071448   BlueFemale  0.007882217       1        L FALSE
#> 99.2...36   -0.066184817   BlueFemale  0.007882217       1        L FALSE
#> 65.1...37   -0.118145516   BlueFemale  0.007882217       1        L FALSE
#> 87...38     -0.025283934   BlueFemale  0.007882217       1        L FALSE
#> 90.1...39   -0.069867755   BlueFemale  0.007882217       1        L FALSE
#> 56...40     -0.070653975   BlueFemale  0.007882217       1        L FALSE
#> 62...41     -0.168746834   BlueFemale  0.007882217       1        L FALSE
#> 64...42     -0.096665410   BlueFemale  0.007882217       1        L FALSE
#> 56.1...43   -0.070653975   BlueFemale  0.007882217       1        L FALSE
#> 56.2...44   -0.070653975   BlueFemale  0.007882217       1        L FALSE
#> 100.1...45  -0.136705852   BlueFemale  0.007882217       1        L FALSE
#> 100.2...46  -0.136705852   BlueFemale  0.007882217       1        L FALSE
#> 57...47     -0.154923986   BlueFemale  0.007882217       1        L FALSE
#> 96.1...48   -0.068970902   BlueFemale  0.007882217       1        L FALSE
#> 79...49     -0.085500462   BlueFemale  0.007882217       1        L FALSE
#> 70...50     -0.219660840   BlueFemale  0.007882217       1        L FALSE
#> 23...51     -0.153904449     BlueMale  0.007882217       1        L FALSE
#> 36...52     -0.216763073     BlueMale  0.007882217       1        L FALSE
#> 5...53      -0.072113336     BlueMale  0.007882217       1        L FALSE
#> 17...54     -0.216875104     BlueMale  0.007882217       1        L FALSE
#> 35...55     -0.285405877     BlueMale  0.007882217       1        L FALSE
#> 16...56     -0.085755434     BlueMale  0.007882217       1        L FALSE
#> 42...57     -0.051693281     BlueMale  0.007882217       1        L FALSE
#> 11...58     -0.204760738     BlueMale  0.007882217       1        L FALSE
#> 25...59     -0.122660690     BlueMale  0.007882217       1        L FALSE
#> 14...60     -0.229817253     BlueMale  0.007882217       1        L FALSE
#> 10...61     -0.162633810     BlueMale  0.007882217       1        L FALSE
#> 48...62     -0.133193011     BlueMale  0.007882217       1        L FALSE
#> 17.1...63   -0.216875104     BlueMale  0.007882217       1        L FALSE
#> 24...64     -0.136489039     BlueMale  0.007882217       1        L FALSE
#> 2...65      -0.074076790     BlueMale  0.007882217       1        L FALSE
#> 28...66     -0.143222637     BlueMale  0.007882217       1        L FALSE
#> 33...67     -0.155267934     BlueMale  0.007882217       1        L FALSE
#> 30...68     -0.124284298     BlueMale  0.007882217       1        L FALSE
#> 17.2...69   -0.216875104     BlueMale  0.007882217       1        L FALSE
#> 45...70     -0.095778161     BlueMale  0.007882217       1        L FALSE
#> 32...71     -0.143636355     BlueMale  0.007882217       1        L FALSE
#> 16.1...72   -0.085755434     BlueMale  0.007882217       1        L FALSE
#> 20...73     -0.108763621     BlueMale  0.007882217       1        L FALSE
#> 9...74      -0.147505544     BlueMale  0.007882217       1        L FALSE
#> 44...75     -0.232766537     BlueMale  0.007882217       1        L FALSE
#> 48.1...76   -0.133193011     BlueMale  0.007882217       1        L FALSE
#> 29...77     -0.133554512     BlueMale  0.007882217       1        L FALSE
#> 17.3...78   -0.216875104     BlueMale  0.007882217       1        L FALSE
#> 50...79     -0.233233169     BlueMale  0.007882217       1        L FALSE
#> 12...80     -0.146279499     BlueMale  0.007882217       1        L FALSE
#> 31...81     -0.052781188     BlueMale  0.007882217       1        L FALSE
#> 6...82      -0.060391382     BlueMale  0.007882217       1        L FALSE
#> 25.1...83   -0.122660690     BlueMale  0.007882217       1        L FALSE
#> 5.1...84    -0.072113336     BlueMale  0.007882217       1        L FALSE
#> 9.1...85    -0.147505544     BlueMale  0.007882217       1        L FALSE
#> 15...86     -0.175230962     BlueMale  0.007882217       1        L FALSE
#> 30.1...87   -0.124284298     BlueMale  0.007882217       1        L FALSE
#> 37...88     -0.151924842     BlueMale  0.007882217       1        L FALSE
#> 9.2...89    -0.147505544     BlueMale  0.007882217       1        L FALSE
#> 44.1...90   -0.232766537     BlueMale  0.007882217       1        L FALSE
#> 36.1...91   -0.216763073     BlueMale  0.007882217       1        L FALSE
#> 39...92     -0.100517047     BlueMale  0.007882217       1        L FALSE
#> 8...93      -0.091459892     BlueMale  0.007882217       1        L FALSE
#> 18...94     -0.192959411     BlueMale  0.007882217       1        L FALSE
#> 34...95     -0.197490390     BlueMale  0.007882217       1        L FALSE
#> 27...96     -0.128513590     BlueMale  0.007882217       1        L FALSE
#> 18.1...97   -0.192959411     BlueMale  0.007882217       1        L FALSE
#> 13...98     -0.122517749     BlueMale  0.007882217       1        L FALSE
#> 3...99      -0.147856043     BlueMale  0.007882217       1        L FALSE
#> 12.1...100  -0.146279499     BlueMale  0.007882217       1        L FALSE
#> 195...101    0.231081900 OrangeFemale  0.007882217       1        R  TRUE
#> 191...102    0.158825227 OrangeFemale  0.007882217       1        R  TRUE
#> 188...103   -0.053571511 OrangeFemale  0.007882217       1        R FALSE
#> 158...104    0.047088873 OrangeFemale  0.007882217       1        R  TRUE
#> 157...105    0.049858804 OrangeFemale  0.007882217       1        R  TRUE
#> 199...106    0.132644044 OrangeFemale  0.007882217       1        R  TRUE
#> 172...107    0.142858175 OrangeFemale  0.007882217       1        R  TRUE
#> 174...108    0.085995392 OrangeFemale  0.007882217       1        R  TRUE
#> 159...109    0.017161184 OrangeFemale  0.007882217       1        R  TRUE
#> 180...110    0.049647143 OrangeFemale  0.007882217       1        R  TRUE
#> 164...111    0.105591418 OrangeFemale  0.007882217       1        R  TRUE
#> 162...112    0.098512822 OrangeFemale  0.007882217       1        R  TRUE
#> 185...113    0.120423149 OrangeFemale  0.007882217       1        R  TRUE
#> 165...114    0.075329385 OrangeFemale  0.007882217       1        R  TRUE
#> 191.1...115  0.158825227 OrangeFemale  0.007882217       1        R  TRUE
#> 155...116    0.119801525 OrangeFemale  0.007882217       1        R  TRUE
#> 192...117    0.221620632 OrangeFemale  0.007882217       1        R  TRUE
#> 172.1...118  0.142858175 OrangeFemale  0.007882217       1        R  TRUE
#> 188.1...119 -0.053571511 OrangeFemale  0.007882217       1        R FALSE
#> 184...120    0.087560933 OrangeFemale  0.007882217       1        R  TRUE
#> 173...121    0.021650600 OrangeFemale  0.007882217       1        R  TRUE
#> 182...122    0.142375389 OrangeFemale  0.007882217       1        R  TRUE
#> 151...123    0.113683000 OrangeFemale  0.007882217       1        R  TRUE
#> 161...124    0.211824321 OrangeFemale  0.007882217       1        R  TRUE
#> 186...125    0.262792291 OrangeFemale  0.007882217       1        R  TRUE
#> 161.1...126  0.211824321 OrangeFemale  0.007882217       1        R  TRUE
#> 191.2...127  0.158825227 OrangeFemale  0.007882217       1        R  TRUE
#> 166...128    0.012693424 OrangeFemale  0.007882217       1        R  TRUE
#> 185.1...129  0.120423149 OrangeFemale  0.007882217       1        R  TRUE
#> 197...130    0.191369252 OrangeFemale  0.007882217       1        R  TRUE
#> 156...131    0.048298763 OrangeFemale  0.007882217       1        R  TRUE
#> 157.1...132  0.049858804 OrangeFemale  0.007882217       1        R  TRUE
#> 184.1...133  0.087560933 OrangeFemale  0.007882217       1        R  TRUE
#> 172.2...134  0.142858175 OrangeFemale  0.007882217       1        R  TRUE
#> 192.1...135  0.221620632 OrangeFemale  0.007882217       1        R  TRUE
#> 199.1...136  0.132644044 OrangeFemale  0.007882217       1        R  TRUE
#> 154...137    0.123377933 OrangeFemale  0.007882217       1        R  TRUE
#> 173.1...138  0.021650600 OrangeFemale  0.007882217       1        R  TRUE
#> 186.1...139  0.262792291 OrangeFemale  0.007882217       1        R  TRUE
#> 155.1...140  0.119801525 OrangeFemale  0.007882217       1        R  TRUE
#> 200...141    0.094831282 OrangeFemale  0.007882217       1        R  TRUE
#> 163...142    0.054183623 OrangeFemale  0.007882217       1        R  TRUE
#> 153...143    0.103038649 OrangeFemale  0.007882217       1        R  TRUE
#> 185.2...144  0.120423149 OrangeFemale  0.007882217       1        R  TRUE
#> 190...145    0.265716165 OrangeFemale  0.007882217       1        R  TRUE
#> 168...146    0.009360985 OrangeFemale  0.007882217       1        R  TRUE
#> 169...147    0.047917006 OrangeFemale  0.007882217       1        R  TRUE
#> 165.1...148  0.075329385 OrangeFemale  0.007882217       1        R  TRUE
#> 154.1...149  0.123377933 OrangeFemale  0.007882217       1        R  TRUE
#> 169.1...150  0.047917006 OrangeFemale  0.007882217       1        R  TRUE
#> 109...151    0.122980021   OrangeMale  0.007882217       1        R  TRUE
#> 127...152    0.199651889   OrangeMale  0.007882217       1        R  TRUE
#> 126...153    0.133672835   OrangeMale  0.007882217       1        R  TRUE
#> 148...154    0.221891408   OrangeMale  0.007882217       1        R  TRUE
#> 118...155    0.007158714   OrangeMale  0.007882217       1        R FALSE
#> 124...156    0.219401858   OrangeMale  0.007882217       1        R  TRUE
#> 105...157    0.145415047   OrangeMale  0.007882217       1        R  TRUE
#> 144...158    0.313982924   OrangeMale  0.007882217       1        R  TRUE
#> 120...159    0.216706148   OrangeMale  0.007882217       1        R  TRUE
#> 137...160    0.281242343   OrangeMale  0.007882217       1        R  TRUE
#> 109.1...161  0.122980021   OrangeMale  0.007882217       1        R  TRUE
#> 123...162    0.228268658   OrangeMale  0.007882217       1        R  TRUE
#> 139...163    0.205844634   OrangeMale  0.007882217       1        R  TRUE
#> 104...164    0.140129807   OrangeMale  0.007882217       1        R  TRUE
#> 141...165    0.177057768   OrangeMale  0.007882217       1        R  TRUE
#> 116...166    0.152350702   OrangeMale  0.007882217       1        R  TRUE
#> 119...167    0.148126911   OrangeMale  0.007882217       1        R  TRUE
#> 116.1...168  0.152350702   OrangeMale  0.007882217       1        R  TRUE
#> 102...169    0.120188435   OrangeMale  0.007882217       1        R  TRUE
#> 126.1...170  0.133672835   OrangeMale  0.007882217       1        R  TRUE
#> 134...171    0.233575553   OrangeMale  0.007882217       1        R  TRUE
#> 146...172    0.246289888   OrangeMale  0.007882217       1        R  TRUE
#> 102.1...173  0.120188435   OrangeMale  0.007882217       1        R  TRUE
#> 150...174    0.205717848   OrangeMale  0.007882217       1        R  TRUE
#> 116.2...175  0.152350702   OrangeMale  0.007882217       1        R  TRUE
#> 147...176    0.207086833   OrangeMale  0.007882217       1        R  TRUE
#> 135...177    0.200713338   OrangeMale  0.007882217       1        R  TRUE
#> 123.1...178  0.228268658   OrangeMale  0.007882217       1        R  TRUE
#> 141.1...179  0.177057768   OrangeMale  0.007882217       1        R  TRUE
#> 102.2...180  0.120188435   OrangeMale  0.007882217       1        R  TRUE
#> 137.1...181  0.281242343   OrangeMale  0.007882217       1        R  TRUE
#> 118.1...182  0.007158714   OrangeMale  0.007882217       1        R FALSE
#> 142...183    0.315680753   OrangeMale  0.007882217       1        R  TRUE
#> 127.1...184  0.199651889   OrangeMale  0.007882217       1        R  TRUE
#> 114...185    0.182777331   OrangeMale  0.007882217       1        R  TRUE
#> 141.2...186  0.177057768   OrangeMale  0.007882217       1        R  TRUE
#> 105.1...187  0.145415047   OrangeMale  0.007882217       1        R  TRUE
#> 142.1...188  0.315680753   OrangeMale  0.007882217       1        R  TRUE
#> 114.1...189  0.182777331   OrangeMale  0.007882217       1        R  TRUE
#> 138...190    0.215199722   OrangeMale  0.007882217       1        R  TRUE
#> 149...191    0.242241346   OrangeMale  0.007882217       1        R  TRUE
#> 105.2...192  0.145415047   OrangeMale  0.007882217       1        R  TRUE
#> 143...193    0.220829959   OrangeMale  0.007882217       1        R  TRUE
#> 141.3...194  0.177057768   OrangeMale  0.007882217       1        R  TRUE
#> 120.1...195  0.216706148   OrangeMale  0.007882217       1        R  TRUE
#> 117...196    0.206045642   OrangeMale  0.007882217       1        R  TRUE
#> 111...197    0.147803220   OrangeMale  0.007882217       1        R  TRUE
#> 129...198    0.222824672   OrangeMale  0.007882217       1        R  TRUE
#> 145...199    0.280064761   OrangeMale  0.007882217       1        R  TRUE
#> 127.2...200  0.199651889   OrangeMale  0.007882217       1        R  TRUE
#> 100...201   -0.175091888   BlueFemale -0.162968406       2        R FALSE
#> 84...202    -0.108167149   BlueFemale -0.162968406       2        R  TRUE
#> 71...203    -0.015382255   BlueFemale -0.162968406       2        R  TRUE
#> 67...204    -0.135408253   BlueFemale -0.162968406       2        R  TRUE
#> 92...205    -0.239067949   BlueFemale -0.162968406       2        R FALSE
#> 58...206    -0.066145149   BlueFemale -0.162968406       2        R  TRUE
#> 60...207    -0.161079605   BlueFemale -0.162968406       2        R  TRUE
#> 95...208    -0.351141033   BlueFemale -0.162968406       2        R FALSE
#> 66...209    -0.087368031   BlueFemale -0.162968406       2        R  TRUE
#> 69...210    -0.027668014   BlueFemale -0.162968406       2        R  TRUE
#> 90...211    -0.090319694   BlueFemale -0.162968406       2        R  TRUE
#> 60.1...212  -0.161079605   BlueFemale -0.162968406       2        R  TRUE
#> 51...213     0.004005905   BlueFemale -0.162968406       2        R  TRUE
#> 71.1...214  -0.015382255   BlueFemale -0.162968406       2        R  TRUE
#> 75...215    -0.290433730   BlueFemale -0.162968406       2        R FALSE
#> 72...216    -0.159450067   BlueFemale -0.162968406       2        R  TRUE
#> 68...217    -0.095281488   BlueFemale -0.162968406       2        R  TRUE
#> 94...218    -0.115613436   BlueFemale -0.162968406       2        R  TRUE
#> 73...219    -0.158244293   BlueFemale -0.162968406       2        R  TRUE
#> 77...220    -0.174071207   BlueFemale -0.162968406       2        R FALSE
#> 99...221    -0.125560830   BlueFemale -0.162968406       2        R  TRUE
#> 89...222    -0.162696475   BlueFemale -0.162968406       2        R  TRUE
#> 61...223    -0.270002539   BlueFemale -0.162968406       2        R FALSE
#> 66.1...224  -0.087368031   BlueFemale -0.162968406       2        R  TRUE
#> 99.1...225  -0.125560830   BlueFemale -0.162968406       2        R  TRUE
#> 80...226    -0.163186723   BlueFemale -0.162968406       2        R FALSE
#> 89.1...227  -0.162696475   BlueFemale -0.162968406       2        R  TRUE
#> 69.1...228  -0.027668014   BlueFemale -0.162968406       2        R  TRUE
#> 89.2...229  -0.162696475   BlueFemale -0.162968406       2        R  TRUE
#> 77.1...230  -0.174071207   BlueFemale -0.162968406       2        R FALSE
#> 96...231    -0.066238422   BlueFemale -0.162968406       2        R  TRUE
#> 60.2...232  -0.161079605   BlueFemale -0.162968406       2        R  TRUE
#> 65...233    -0.092131339   BlueFemale -0.162968406       2        R  TRUE
#> 55...234    -0.088477546   BlueFemale -0.162968406       2        R  TRUE
#> 61.1...235  -0.270002539   BlueFemale -0.162968406       2        R FALSE
#> 99.2...236  -0.125560830   BlueFemale -0.162968406       2        R  TRUE
#> 65.1...237  -0.092131339   BlueFemale -0.162968406       2        R  TRUE
#> 87...238    -0.082104794   BlueFemale -0.162968406       2        R  TRUE
#> 90.1...239  -0.090319694   BlueFemale -0.162968406       2        R  TRUE
#> 56...240    -0.027571756   BlueFemale -0.162968406       2        R  TRUE
#> 62...241    -0.145945139   BlueFemale -0.162968406       2        R  TRUE
#> 64...242    -0.090925565   BlueFemale -0.162968406       2        R  TRUE
#> 56.1...243  -0.027571756   BlueFemale -0.162968406       2        R  TRUE
#> 56.2...244  -0.027571756   BlueFemale -0.162968406       2        R  TRUE
#> 100.1...245 -0.175091888   BlueFemale -0.162968406       2        R FALSE
#> 100.2...246 -0.175091888   BlueFemale -0.162968406       2        R FALSE
#> 57...247    -0.140873691   BlueFemale -0.162968406       2        R  TRUE
#> 96.1...248  -0.066238422   BlueFemale -0.162968406       2        R  TRUE
#> 79...249    -0.095400824   BlueFemale -0.162968406       2        R  TRUE
#> 70...250    -0.194594944   BlueFemale -0.162968406       2        R FALSE
#> 23...251    -0.178735271     BlueMale -0.162968406       2        L FALSE
#> 36...252    -0.298450143     BlueMale -0.162968406       2        L FALSE
#> 5...253     -0.057711669     BlueMale -0.162968406       2        L  TRUE
#> 17...254    -0.281291422     BlueMale -0.162968406       2        L FALSE
#> 35...255    -0.456665227     BlueMale -0.162968406       2        L FALSE
#> 16...256    -0.079438194     BlueMale -0.162968406       2        L  TRUE
#> 42...257    -0.159152336     BlueMale -0.162968406       2        L  TRUE
#> 11...258    -0.266391916     BlueMale -0.162968406       2        L FALSE
#> 25...259    -0.177529497     BlueMale -0.162968406       2        L FALSE
#> 14...260    -0.263934215     BlueMale -0.162968406       2        L FALSE
#> 10...261    -0.178549451     BlueMale -0.162968406       2        L FALSE
#> 48...262    -0.227789476     BlueMale -0.162968406       2        L FALSE
#> 17.1...263  -0.281291422     BlueMale -0.162968406       2        L FALSE
#> 24...264    -0.197019886     BlueMale -0.162968406       2        L FALSE
#> 2...265     -0.039036049     BlueMale -0.162968406       2        L  TRUE
#> 28...266    -0.272682533     BlueMale -0.162968406       2        L FALSE
#> 33...267    -0.231244054     BlueMale -0.162968406       2        L FALSE
#> 30...268    -0.154872579     BlueMale -0.162968406       2        L  TRUE
#> 17.2...269  -0.281291422     BlueMale -0.162968406       2        L FALSE
#> 45...270    -0.181882435     BlueMale -0.162968406       2        L FALSE
#> 32...271    -0.214406869     BlueMale -0.162968406       2        L FALSE
#> 16.1...272  -0.079438194     BlueMale -0.162968406       2        L  TRUE
#> 20...273    -0.136190262     BlueMale -0.162968406       2        L  TRUE
#> 9...274     -0.152026858     BlueMale -0.162968406       2        L  TRUE
#> 44...275    -0.350770121     BlueMale -0.162968406       2        L FALSE
#> 48.1...276  -0.227789476     BlueMale -0.162968406       2        L FALSE
#> 29...277    -0.190921787     BlueMale -0.162968406       2        L FALSE
#> 17.3...278  -0.281291422     BlueMale -0.162968406       2        L FALSE
#> 50...279    -0.368525031     BlueMale -0.162968406       2        L FALSE
#> 12...280    -0.157445905     BlueMale -0.162968406       2        L  TRUE
#> 31...281    -0.065960057     BlueMale -0.162968406       2        L  TRUE
#> 6...282     -0.045644489     BlueMale -0.162968406       2        L  TRUE
#> 25.1...283  -0.177529497     BlueMale -0.162968406       2        L FALSE
#> 5.1...284   -0.057711669     BlueMale -0.162968406       2        L  TRUE
#> 9.1...285   -0.152026858     BlueMale -0.162968406       2        L  TRUE
#> 15...286    -0.216414016     BlueMale -0.162968406       2        L FALSE
#> 30.1...287  -0.154872579     BlueMale -0.162968406       2        L  TRUE
#> 37...288    -0.245563751     BlueMale -0.162968406       2        L FALSE
#> 9.2...289   -0.152026858     BlueMale -0.162968406       2        L  TRUE
#> 44.1...290  -0.350770121     BlueMale -0.162968406       2        L FALSE
#> 36.1...291  -0.298450143     BlueMale -0.162968406       2        L FALSE
#> 39...292    -0.162898674     BlueMale -0.162968406       2        L  TRUE
#> 8...293     -0.081481814     BlueMale -0.162968406       2        L  TRUE
#> 18...294    -0.252357283     BlueMale -0.162968406       2        L FALSE
#> 34...295    -0.255113443     BlueMale -0.162968406       2        L FALSE
#> 27...296    -0.191630614     BlueMale -0.162968406       2        L FALSE
#> 18.1...297  -0.252357283     BlueMale -0.162968406       2        L FALSE
#> 13...298    -0.165660805     BlueMale -0.162968406       2        L FALSE
#> 3...299     -0.118339093     BlueMale -0.162968406       2        L  TRUE
#> 12.1...300  -0.157445905     BlueMale -0.162968406       2        L  TRUE
#> 195...301   -0.657684908 OrangeFemale -0.086472974       3        L FALSE
#> 191...302   -0.312964701 OrangeFemale -0.086472974       3        L FALSE
#> 188...303   -0.678603025 OrangeFemale -0.086472974       3        L FALSE
#> 158...304   -0.227506216 OrangeFemale -0.086472974       3        L FALSE
#> 157...305   -0.353601411 OrangeFemale -0.086472974       3        L FALSE
#> 199...306   -0.280106152 OrangeFemale -0.086472974       3        L FALSE
#> 172...307   -0.270559869 OrangeFemale -0.086472974       3        L FALSE
#> 174...308   -0.186554413 OrangeFemale -0.086472974       3        L FALSE
#> 159...309   -0.387126677 OrangeFemale -0.086472974       3        L FALSE
#> 180...310   -0.078544919 OrangeFemale -0.086472974       3        L  TRUE
#> 164...311   -0.558814324 OrangeFemale -0.086472974       3        L FALSE
#> 162...312   -0.332860896 OrangeFemale -0.086472974       3        L FALSE
#> 185...313   -0.526669763 OrangeFemale -0.086472974       3        L FALSE
#> 165...314   -0.387133836 OrangeFemale -0.086472974       3        L FALSE
#> 191.1...315 -0.312964701 OrangeFemale -0.086472974       3        L FALSE
#> 155...316   -0.160162065 OrangeFemale -0.086472974       3        L FALSE
#> 192...317   -0.635987638 OrangeFemale -0.086472974       3        L FALSE
#> 172.1...318 -0.270559869 OrangeFemale -0.086472974       3        L FALSE
#> 188.1...319 -0.678603025 OrangeFemale -0.086472974       3        L FALSE
#> 184...320   -0.463709177 OrangeFemale -0.086472974       3        L FALSE
#> 173...321   -0.475232818 OrangeFemale -0.086472974       3        L FALSE
#> 182...322   -0.379348517 OrangeFemale -0.086472974       3        L FALSE
#> 151...323   -0.146732071 OrangeFemale -0.086472974       3        L FALSE
#> 161...324   -0.131382725 OrangeFemale -0.086472974       3        L FALSE
#> 186...325   -0.560253037 OrangeFemale -0.086472974       3        L FALSE
#> 161.1...326 -0.131382725 OrangeFemale -0.086472974       3        L FALSE
#> 191.2...327 -0.312964701 OrangeFemale -0.086472974       3        L FALSE
#> 166...328   -0.286869549 OrangeFemale -0.086472974       3        L FALSE
#> 185.1...329 -0.526669763 OrangeFemale -0.086472974       3        L FALSE
#> 197...330   -0.376287650 OrangeFemale -0.086472974       3        L FALSE
#> 156...331   -0.173055674 OrangeFemale -0.086472974       3        L FALSE
#> 157.1...332 -0.353601411 OrangeFemale -0.086472974       3        L FALSE
#> 184.1...333 -0.463709177 OrangeFemale -0.086472974       3        L FALSE
#> 172.2...334 -0.270559869 OrangeFemale -0.086472974       3        L FALSE
#> 192.1...335 -0.635987638 OrangeFemale -0.086472974       3        L FALSE
#> 199.1...336 -0.280106152 OrangeFemale -0.086472974       3        L FALSE
#> 154...337   -0.331664951 OrangeFemale -0.086472974       3        L FALSE
#> 173.1...338 -0.475232818 OrangeFemale -0.086472974       3        L FALSE
#> 186.1...339 -0.560253037 OrangeFemale -0.086472974       3        L FALSE
#> 155.1...340 -0.160162065 OrangeFemale -0.086472974       3        L FALSE
#> 200...341   -0.842675796 OrangeFemale -0.086472974       3        L FALSE
#> 163...342   -0.462803270 OrangeFemale -0.086472974       3        L FALSE
#> 153...343    0.016424057 OrangeFemale -0.086472974       3        L  TRUE
#> 185.2...344 -0.526669763 OrangeFemale -0.086472974       3        L FALSE
#> 190...345   -0.664955317 OrangeFemale -0.086472974       3        L FALSE
#> 168...346   -0.437447160 OrangeFemale -0.086472974       3        L FALSE
#> 169...347   -0.408421474 OrangeFemale -0.086472974       3        L FALSE
#> 165.1...348 -0.387133836 OrangeFemale -0.086472974       3        L FALSE
#> 154.1...349 -0.331664951 OrangeFemale -0.086472974       3        L FALSE
#> 169.1...350 -0.408421474 OrangeFemale -0.086472974       3        L FALSE
#> 109...351    0.204359797   OrangeMale -0.086472974       3        R  TRUE
#> 127...352    0.240228364   OrangeMale -0.086472974       3        R  TRUE
#> 126...353    0.248683980   OrangeMale -0.086472974       3        R  TRUE
#> 148...354    0.205206377   OrangeMale -0.086472974       3        R  TRUE
#> 118...355    0.178286120   OrangeMale -0.086472974       3        R  TRUE
#> 124...356    0.182648210   OrangeMale -0.086472974       3        R  TRUE
#> 105...357    0.109001144   OrangeMale -0.086472974       3        R  TRUE
#> 144...358    0.301163002   OrangeMale -0.086472974       3        R  TRUE
#> 120...359    0.132581233   OrangeMale -0.086472974       3        R  TRUE
#> 137...360    0.198479512   OrangeMale -0.086472974       3        R  TRUE
#> 109.1...361  0.204359797   OrangeMale -0.086472974       3        R  TRUE
#> 123...362   -0.026629597   OrangeMale -0.086472974       3        R  TRUE
#> 139...363    0.168623822   OrangeMale -0.086472974       3        R  TRUE
#> 104...364    0.125622337   OrangeMale -0.086472974       3        R  TRUE
#> 141...365    0.394689687   OrangeMale -0.086472974       3        R  TRUE
#> 116...366    0.137254835   OrangeMale -0.086472974       3        R  TRUE
#> 119...367    0.120285596   OrangeMale -0.086472974       3        R  TRUE
#> 116.1...368  0.137254835   OrangeMale -0.086472974       3        R  TRUE
#> 102...369    0.142772758   OrangeMale -0.086472974       3        R  TRUE
#> 126.1...370  0.248683980   OrangeMale -0.086472974       3        R  TRUE
#> 134...371    0.289943716   OrangeMale -0.086472974       3        R  TRUE
#> 146...372    0.381440874   OrangeMale -0.086472974       3        R  TRUE
#> 102.1...373  0.142772758   OrangeMale -0.086472974       3        R  TRUE
#> 150...374    0.501862019   OrangeMale -0.086472974       3        R  TRUE
#> 116.2...375  0.137254835   OrangeMale -0.086472974       3        R  TRUE
#> 147...376    0.351117544   OrangeMale -0.086472974       3        R  TRUE
#> 135...377    0.265051667   OrangeMale -0.086472974       3        R  TRUE
#> 123.1...378 -0.026629597   OrangeMale -0.086472974       3        R  TRUE
#> 141.1...379  0.394689687   OrangeMale -0.086472974       3        R  TRUE
#> 102.2...380  0.142772758   OrangeMale -0.086472974       3        R  TRUE
#> 137.1...381  0.198479512   OrangeMale -0.086472974       3        R  TRUE
#> 118.1...382  0.178286120   OrangeMale -0.086472974       3        R  TRUE
#> 142...383    0.356439968   OrangeMale -0.086472974       3        R  TRUE
#> 127.1...384  0.240228364   OrangeMale -0.086472974       3        R  TRUE
#> 114...385    0.220557041   OrangeMale -0.086472974       3        R  TRUE
#> 141.2...386  0.394689687   OrangeMale -0.086472974       3        R  TRUE
#> 105.1...387  0.109001144   OrangeMale -0.086472974       3        R  TRUE
#> 142.1...388  0.356439968   OrangeMale -0.086472974       3        R  TRUE
#> 114.1...389  0.220557041   OrangeMale -0.086472974       3        R  TRUE
#> 138...390    0.264580448   OrangeMale -0.086472974       3        R  TRUE
#> 149...391    0.120900145   OrangeMale -0.086472974       3        R  TRUE
#> 105.2...392  0.109001144   OrangeMale -0.086472974       3        R  TRUE
#> 143...393    0.239086848   OrangeMale -0.086472974       3        R  TRUE
#> 141.3...394  0.394689687   OrangeMale -0.086472974       3        R  TRUE
#> 120.1...395  0.132581233   OrangeMale -0.086472974       3        R  TRUE
#> 117...396    0.069685812   OrangeMale -0.086472974       3        R  TRUE
#> 111...397    0.010924030   OrangeMale -0.086472974       3        R  TRUE
#> 129...398    0.177913020   OrangeMale -0.086472974       3        R  TRUE
#> 145...399    0.330543893   OrangeMale -0.086472974       3        R  TRUE
#> 127.2...400  0.240228364   OrangeMale -0.086472974       3        R  TRUE