Obtain the permuted importance variable measure

Obtain the permuted importance variable measure

Usage

permute_importance(ppf)

Arguments

ppf

is a PPforest object

Value

A data frame with permuted importance measures, imp is the permuted importance measure defined in Brieman paper, imp2 is the permuted importance measure defined in randomForest package, the standard deviation (sd.im and sd.imp2) for each measure is computed and the also the standardized mesure.

Examples

pprf.crab <- PPforest(data = crab, y = 'Type',
std = 'min-max', size.tr = 1, m = 100, size.p = .4, PPmethod = 'LDA', parallel = TRUE, core = 2)
permute_importance(ppf = pprf.crab) 
#>   nm   imp   sd.imp      imp2   sd.imp2 imp2.std  imp.std
#> 1 CW 12.33 10.18620 0.1683782 0.1369413 1.229565 1.210461
#> 2 CL 13.53 11.46664 0.1863579 0.1566416 1.189709 1.179944
#> 3 BD 15.38 12.76420 0.2105262 0.1738368 1.211056 1.204932
#> 4 FL 17.02 13.37190 0.2329609 0.1807638 1.288759 1.272818
#> 5 RW 17.32 11.24734 0.2363852 0.1525993 1.549058 1.539920