Global importance measure for a PPforest object

Global importance measure for a PPforest object

Usage

ppf_global_imp(data, class, ppf)

Arguments

data

Data frame with the complete data set.

class

A character with the name of the class variable.

ppf

is a PPforest object

Value

Data frame with the global importance measure

Examples

#crab data set with all the observations used as training

pprf.crab <- PPforest(data = crab, std = TRUE, class = 'Type',
 size.tr = 1, m = 200, size.p = .5, PPmethod = 'LDA', parallel = TRUE, cores = 2)
 
ppf_global_imp(data = crab, class = 'Type', pprf.crab) 
#> # A tibble: 5 × 2
#>   variable  mean
#>   <fct>    <dbl>
#> 1 CW       0.389
#> 2 CL       0.369
#> 3 RW       0.298
#> 4 BD       0.247
#> 5 FL       0.229