Data structure with the projected and boundary by node and class.

Data structure with the projected and boundary by node and class.

Usage

node_data(ppf, tr, Rule = 1)

Arguments

ppf

is a PPforest object

tr

numerical value to identify a tree

Rule

split rule 1:mean of two group means, 2:weighted mean, 3: mean of max(left group) and min(right group), 4: weighted mean of max(left group) and min(right group)

Value

Data frame with projected data for each class and node id and the boundaries

Examples

#crab data set with all the observations used as training

pprf.crab <- PPforest(data = crab, std =TRUE, class = 'Type',
 size.tr = 1, m = 200, size.p = .5, PPmethod = 'LDA')
node_data(ppf = pprf.crab, tr = 1) 
#>         proj.data Class         cut node.id LR.class   Dir
#> 1   -0.0495768213     1  0.03547448       1        L FALSE
#> 2   -0.0359435011     1  0.03547448       1        L FALSE
#> 3   -0.1100147873     1  0.03547448       1        L FALSE
#> 4   -0.0495768213     1  0.03547448       1        L FALSE
#> 5   -0.0543366596     1  0.03547448       1        L FALSE
#> 6   -0.0873789996     1  0.03547448       1        L FALSE
#> 7   -0.0309852962     1  0.03547448       1        L FALSE
#> 8   -0.0955751011     1  0.03547448       1        L FALSE
#> 9   -0.0372703473     1  0.03547448       1        L FALSE
#> 10  -0.0556364394     1  0.03547448       1        L FALSE
#> 11  -0.0628172205     1  0.03547448       1        L FALSE
#> 12  -0.0114046316     1  0.03547448       1        L FALSE
#> 13  -0.0372703473     1  0.03547448       1        L FALSE
#> 14  -0.1085933387     1  0.03547448       1        L FALSE
#> 15  -0.0395493308     1  0.03547448       1        L FALSE
#> 16  -0.0862343861     1  0.03547448       1        L FALSE
#> 17  -0.0516803480     1  0.03547448       1        L FALSE
#> 18  -0.0934045792     1  0.03547448       1        L FALSE
#> 19  -0.1281070080     1  0.03547448       1        L FALSE
#> 20  -0.1281070080     1  0.03547448       1        L FALSE
#> 21  -0.0309852962     1  0.03547448       1        L FALSE
#> 22  -0.0395493308     1  0.03547448       1        L FALSE
#> 23  -0.1157708485     1  0.03547448       1        L FALSE
#> 24  -0.0873789996     1  0.03547448       1        L FALSE
#> 25  -0.1085933387     1  0.03547448       1        L FALSE
#> 26  -0.1461839432     1  0.03547448       1        L FALSE
#> 27  -0.0844153189     1  0.03547448       1        L FALSE
#> 28  -0.0934045792     1  0.03547448       1        L FALSE
#> 29  -0.0873789996     1  0.03547448       1        L FALSE
#> 30  -0.0309852962     1  0.03547448       1        L FALSE
#> 31  -0.0302709136     1  0.03547448       1        L FALSE
#> 32  -0.0512072245     1  0.03547448       1        L FALSE
#> 33  -0.0556364394     1  0.03547448       1        L FALSE
#> 34  -0.0436978985     1  0.03547448       1        L FALSE
#> 35  -0.0838146362     1  0.03547448       1        L FALSE
#> 36  -0.0903112610     1  0.03547448       1        L FALSE
#> 37  -0.0385281199     1  0.03547448       1        L FALSE
#> 38  -0.1100147873     1  0.03547448       1        L FALSE
#> 39  -0.1100147873     1  0.03547448       1        L FALSE
#> 40  -0.0495768213     1  0.03547448       1        L FALSE
#> 41  -0.0934045792     1  0.03547448       1        L FALSE
#> 42  -0.1281070080     1  0.03547448       1        L FALSE
#> 43  -0.0903112610     1  0.03547448       1        L FALSE
#> 44   0.0129955498     1  0.03547448       1        L FALSE
#> 45  -0.1085933387     1  0.03547448       1        L FALSE
#> 46  -0.0858570581     1  0.03547448       1        L FALSE
#> 47  -0.1343373855     1  0.03547448       1        L FALSE
#> 48  -0.0709277701     1  0.03547448       1        L FALSE
#> 49  -0.0543366596     1  0.03547448       1        L FALSE
#> 50  -0.1105942941     1  0.03547448       1        L FALSE
#> 51  -0.0279607071     2  0.03547448       1        L FALSE
#> 52  -0.0051595095     2  0.03547448       1        L FALSE
#> 53   0.0230080993     2  0.03547448       1        L FALSE
#> 54   0.0278029726     2  0.03547448       1        L FALSE
#> 55   0.0393652673     2  0.03547448       1        L  TRUE
#> 56  -0.0164867799     2  0.03547448       1        L FALSE
#> 57  -0.0201016602     2  0.03547448       1        L FALSE
#> 58  -0.0542569980     2  0.03547448       1        L FALSE
#> 59   0.0055441685     2  0.03547448       1        L FALSE
#> 60  -0.0372705436     2  0.03547448       1        L FALSE
#> 61  -0.0259156661     2  0.03547448       1        L FALSE
#> 62   0.0547586283     2  0.03547448       1        L  TRUE
#> 63  -0.0221917833     2  0.03547448       1        L FALSE
#> 64   0.0117707339     2  0.03547448       1        L FALSE
#> 65  -0.0259156661     2  0.03547448       1        L FALSE
#> 66  -0.0002847784     2  0.03547448       1        L FALSE
#> 67   0.0245315783     2  0.03547448       1        L FALSE
#> 68   0.0048159266     2  0.03547448       1        L FALSE
#> 69  -0.0527876517     2  0.03547448       1        L FALSE
#> 70  -0.0224559522     2  0.03547448       1        L FALSE
#> 71  -0.0278687238     2  0.03547448       1        L FALSE
#> 72  -0.0542569980     2  0.03547448       1        L FALSE
#> 73  -0.0195301221     2  0.03547448       1        L FALSE
#> 74   0.0299964042     2  0.03547448       1        L FALSE
#> 75  -0.0201016602     2  0.03547448       1        L FALSE
#> 76  -0.0059547467     2  0.03547448       1        L FALSE
#> 77   0.0245315783     2  0.03547448       1        L FALSE
#> 78  -0.0147329742     2  0.03547448       1        L FALSE
#> 79  -0.0714451754     2  0.03547448       1        L FALSE
#> 80   0.0140879126     2  0.03547448       1        L FALSE
#> 81  -0.0512806511     2  0.03547448       1        L FALSE
#> 82  -0.0224559522     2  0.03547448       1        L FALSE
#> 83   0.0549032067     2  0.03547448       1        L  TRUE
#> 84  -0.0372705436     2  0.03547448       1        L FALSE
#> 85  -0.0221917833     2  0.03547448       1        L FALSE
#> 86   0.0159128702     2  0.03547448       1        L FALSE
#> 87  -0.0934653395     2  0.03547448       1        L FALSE
#> 88  -0.0527876517     2  0.03547448       1        L FALSE
#> 89   0.0055441685     2  0.03547448       1        L FALSE
#> 90  -0.0224559522     2  0.03547448       1        L FALSE
#> 91  -0.0147329742     2  0.03547448       1        L FALSE
#> 92  -0.0278687238     2  0.03547448       1        L FALSE
#> 93  -0.0201016602     2  0.03547448       1        L FALSE
#> 94   0.0140879126     2  0.03547448       1        L FALSE
#> 95  -0.0044078171     2  0.03547448       1        L FALSE
#> 96   0.0140879126     2  0.03547448       1        L FALSE
#> 97  -0.0094000604     2  0.03547448       1        L FALSE
#> 98  -0.0512806511     2  0.03547448       1        L FALSE
#> 99  -0.0221917833     2  0.03547448       1        L FALSE
#> 100 -0.0044078171     2  0.03547448       1        L FALSE
#> 101  0.0038655231     3  0.03547448       1        L FALSE
#> 102 -0.0483907414     3  0.03547448       1        L FALSE
#> 103 -0.0298530045     3  0.03547448       1        L FALSE
#> 104 -0.0016433884     3  0.03547448       1        L FALSE
#> 105 -0.0139524816     3  0.03547448       1        L FALSE
#> 106  0.0389268344     3  0.03547448       1        L  TRUE
#> 107  0.0188027799     3  0.03547448       1        L FALSE
#> 108 -0.0759788435     3  0.03547448       1        L FALSE
#> 109 -0.0804007418     3  0.03547448       1        L FALSE
#> 110  0.0068227725     3  0.03547448       1        L FALSE
#> 111  0.0069312341     3  0.03547448       1        L FALSE
#> 112  0.0038655231     3  0.03547448       1        L FALSE
#> 113 -0.0162026650     3  0.03547448       1        L FALSE
#> 114  0.0479810117     3  0.03547448       1        L  TRUE
#> 115  0.0016174180     3  0.03547448       1        L FALSE
#> 116 -0.0016433884     3  0.03547448       1        L FALSE
#> 117 -0.0157838706     3  0.03547448       1        L FALSE
#> 118  0.0065336156     3  0.03547448       1        L FALSE
#> 119  0.0255470951     3  0.03547448       1        L FALSE
#> 120  0.0188027799     3  0.03547448       1        L FALSE
#> 121  0.0089703847     3  0.03547448       1        L FALSE
#> 122 -0.0759788435     3  0.03547448       1        L FALSE
#> 123  0.0479810117     3  0.03547448       1        L  TRUE
#> 124  0.0208622209     3  0.03547448       1        L FALSE
#> 125 -0.0079895442     3  0.03547448       1        L FALSE
#> 126 -0.0981000414     3  0.03547448       1        L FALSE
#> 127 -0.1152748153     3  0.03547448       1        L FALSE
#> 128 -0.0139524816     3  0.03547448       1        L FALSE
#> 129  0.0038655231     3  0.03547448       1        L FALSE
#> 130 -0.0216483934     3  0.03547448       1        L FALSE
#> 131  0.0197884148     3  0.03547448       1        L FALSE
#> 132 -0.0016433884     3  0.03547448       1        L FALSE
#> 133 -0.0609969603     3  0.03547448       1        L FALSE
#> 134 -0.0503352896     3  0.03547448       1        L FALSE
#> 135  0.0038655231     3  0.03547448       1        L FALSE
#> 136 -0.0016433884     3  0.03547448       1        L FALSE
#> 137 -0.0263912123     3  0.03547448       1        L FALSE
#> 138 -0.0804007418     3  0.03547448       1        L FALSE
#> 139  0.0196294363     3  0.03547448       1        L FALSE
#> 140 -0.0298530045     3  0.03547448       1        L FALSE
#> 141  0.0606195353     3  0.03547448       1        L  TRUE
#> 142  0.0068227725     3  0.03547448       1        L FALSE
#> 143  0.0089703847     3  0.03547448       1        L FALSE
#> 144 -0.0503352896     3  0.03547448       1        L FALSE
#> 145 -0.0157838706     3  0.03547448       1        L FALSE
#> 146 -0.0079895442     3  0.03547448       1        L FALSE
#> 147  0.0068227725     3  0.03547448       1        L FALSE
#> 148  0.0688300369     3  0.03547448       1        L  TRUE
#> 149 -0.0981000414     3  0.03547448       1        L FALSE
#> 150 -0.0348053189     3  0.03547448       1        L FALSE
#> 151  0.1650389993     4  0.03547448       1        R  TRUE
#> 152  0.1807720341     4  0.03547448       1        R  TRUE
#> 153  0.0595158473     4  0.03547448       1        R  TRUE
#> 154  0.1858157909     4  0.03547448       1        R  TRUE
#> 155  0.0832212845     4  0.03547448       1        R  TRUE
#> 156  0.0914344052     4  0.03547448       1        R  TRUE
#> 157  0.0378077498     4  0.03547448       1        R  TRUE
#> 158  0.1339847521     4  0.03547448       1        R  TRUE
#> 159  0.0872087953     4  0.03547448       1        R  TRUE
#> 160  0.0691382913     4  0.03547448       1        R  TRUE
#> 161  0.1552480706     4  0.03547448       1        R  TRUE
#> 162  0.0901325470     4  0.03547448       1        R  TRUE
#> 163  0.0560419298     4  0.03547448       1        R  TRUE
#> 164  0.1492660357     4  0.03547448       1        R  TRUE
#> 165  0.0914344052     4  0.03547448       1        R  TRUE
#> 166  0.0595158473     4  0.03547448       1        R  TRUE
#> 167  0.0595158473     4  0.03547448       1        R  TRUE
#> 168  0.0707410873     4  0.03547448       1        R  TRUE
#> 169  0.1665959760     4  0.03547448       1        R  TRUE
#> 170  0.0904631703     4  0.03547448       1        R  TRUE
#> 171  0.1492660357     4  0.03547448       1        R  TRUE
#> 172  0.1102336919     4  0.03547448       1        R  TRUE
#> 173  0.0753122615     4  0.03547448       1        R  TRUE
#> 174  0.0944065956     4  0.03547448       1        R  TRUE
#> 175  0.1213599765     4  0.03547448       1        R  TRUE
#> 176  0.1552480706     4  0.03547448       1        R  TRUE
#> 177  0.0753122615     4  0.03547448       1        R  TRUE
#> 178  0.0872087953     4  0.03547448       1        R  TRUE
#> 179  0.0894043052     4  0.03547448       1        R  TRUE
#> 180  0.1120230736     4  0.03547448       1        R  TRUE
#> 181  0.1335935648     4  0.03547448       1        R  TRUE
#> 182  0.0378077498     4  0.03547448       1        R  TRUE
#> 183  0.1027472756     4  0.03547448       1        R  TRUE
#> 184  0.1056960153     4  0.03547448       1        R  TRUE
#> 185  0.1492660357     4  0.03547448       1        R  TRUE
#> 186  0.0914243582     4  0.03547448       1        R  TRUE
#> 187  0.0753122615     4  0.03547448       1        R  TRUE
#> 188  0.0752117687     4  0.03547448       1        R  TRUE
#> 189  0.1335935648     4  0.03547448       1        R  TRUE
#> 190  0.1603987516     4  0.03547448       1        R  TRUE
#> 191  0.0560419298     4  0.03547448       1        R  TRUE
#> 192  0.1603987516     4  0.03547448       1        R  TRUE
#> 193  0.1120400929     4  0.03547448       1        R  TRUE
#> 194  0.1102336919     4  0.03547448       1        R  TRUE
#> 195  0.1335935648     4  0.03547448       1        R  TRUE
#> 196  0.0872087953     4  0.03547448       1        R  TRUE
#> 197  0.0831798181     4  0.03547448       1        R  TRUE
#> 198  0.0378077498     4  0.03547448       1        R  TRUE
#> 199  0.1120230736     4  0.03547448       1        R  TRUE
#> 200  0.0836109343     4  0.03547448       1        R  TRUE
#> 201  0.0794275610     1 -0.01348186       2        R  TRUE
#> 202  0.1918300893     1 -0.01348186       2        R  TRUE
#> 203  0.3411855883     1 -0.01348186       2        R  TRUE
#> 204  0.0794275610     1 -0.01348186       2        R  TRUE
#> 205  0.2702726227     1 -0.01348186       2        R  TRUE
#> 206  0.1685408373     1 -0.01348186       2        R  TRUE
#> 207  0.2679864667     1 -0.01348186       2        R  TRUE
#> 208  0.2839722440     1 -0.01348186       2        R  TRUE
#> 209  0.2718086942     1 -0.01348186       2        R  TRUE
#> 210  0.1315446020     1 -0.01348186       2        R  TRUE
#> 211  0.1778505211     1 -0.01348186       2        R  TRUE
#> 212  0.0021225840     1 -0.01348186       2        R  TRUE
#> 213  0.2718086942     1 -0.01348186       2        R  TRUE
#> 214  0.2444099058     1 -0.01348186       2        R  TRUE
#> 215  0.1627476667     1 -0.01348186       2        R  TRUE
#> 216  0.5102664264     1 -0.01348186       2        R  TRUE
#> 217  0.0383507836     1 -0.01348186       2        R  TRUE
#> 218  0.3900612914     1 -0.01348186       2        R  TRUE
#> 219  0.5752860585     1 -0.01348186       2        R  TRUE
#> 220  0.5752860585     1 -0.01348186       2        R  TRUE
#> 221  0.2679864667     1 -0.01348186       2        R  TRUE
#> 222  0.1627476667     1 -0.01348186       2        R  TRUE
#> 223  0.4769959367     1 -0.01348186       2        R  TRUE
#> 224  0.1685408373     1 -0.01348186       2        R  TRUE
#> 225  0.2444099058     1 -0.01348186       2        R  TRUE
#> 226  0.3620489378     1 -0.01348186       2        R  TRUE
#> 227  0.3434896956     1 -0.01348186       2        R  TRUE
#> 228  0.3900612914     1 -0.01348186       2        R  TRUE
#> 229  0.1685408373     1 -0.01348186       2        R  TRUE
#> 230  0.2679864667     1 -0.01348186       2        R  TRUE
#> 231  0.0432030425     1 -0.01348186       2        R  TRUE
#> 232  0.0757128585     1 -0.01348186       2        R  TRUE
#> 233  0.1315446020     1 -0.01348186       2        R  TRUE
#> 234  0.1032555285     1 -0.01348186       2        R  TRUE
#> 235  0.2596635498     1 -0.01348186       2        R  TRUE
#> 236  0.2349600215     1 -0.01348186       2        R  TRUE
#> 237  0.2079376618     1 -0.01348186       2        R  TRUE
#> 238  0.3411855883     1 -0.01348186       2        R  TRUE
#> 239  0.3411855883     1 -0.01348186       2        R  TRUE
#> 240  0.0794275610     1 -0.01348186       2        R  TRUE
#> 241  0.3900612914     1 -0.01348186       2        R  TRUE
#> 242  0.5752860585     1 -0.01348186       2        R  TRUE
#> 243  0.2349600215     1 -0.01348186       2        R  TRUE
#> 244 -0.0367829246     1 -0.01348186       2        R FALSE
#> 245  0.2444099058     1 -0.01348186       2        R  TRUE
#> 246  0.1692835597     1 -0.01348186       2        R  TRUE
#> 247  0.2785923127     1 -0.01348186       2        R  TRUE
#> 248  0.0972367188     1 -0.01348186       2        R  TRUE
#> 249  0.2702726227     1 -0.01348186       2        R  TRUE
#> 250  0.0919896259     1 -0.01348186       2        R  TRUE
#> 251 -0.4841663705     2 -0.01348186       2        L FALSE
#> 252 -0.2854185391     2 -0.01348186       2        L FALSE
#> 253 -0.2470264328     2 -0.01348186       2        L FALSE
#> 254 -0.7511245450     2 -0.01348186       2        L FALSE
#> 255 -0.1646361956     2 -0.01348186       2        L FALSE
#> 256 -0.5074666656     2 -0.01348186       2        L FALSE
#> 257 -0.7157972196     2 -0.01348186       2        L FALSE
#> 258 -0.0400880688     2 -0.01348186       2        L FALSE
#> 259 -0.1785614560     2 -0.01348186       2        L FALSE
#> 260 -0.6059897886     2 -0.01348186       2        L FALSE
#> 261 -0.4657510097     2 -0.01348186       2        L FALSE
#> 262 -0.3785690930     2 -0.01348186       2        L FALSE
#> 263 -0.1175511975     2 -0.01348186       2        L FALSE
#> 264 -0.2050451002     2 -0.01348186       2        L FALSE
#> 265 -0.4657510097     2 -0.01348186       2        L FALSE
#> 266 -0.2220788601     2 -0.01348186       2        L FALSE
#> 267 -0.5920428959     2 -0.01348186       2        L FALSE
#> 268  0.0264529105     2 -0.01348186       2        L  TRUE
#> 269  0.0523087195     2 -0.01348186       2        L  TRUE
#> 270 -0.0544228875     2 -0.01348186       2        L FALSE
#> 271 -0.4289988188     2 -0.01348186       2        L FALSE
#> 272 -0.0400880688     2 -0.01348186       2        L FALSE
#> 273 -0.3720548323     2 -0.01348186       2        L FALSE
#> 274 -0.4460183085     2 -0.01348186       2        L FALSE
#> 275 -0.7157972196     2 -0.01348186       2        L FALSE
#> 276 -0.1454164426     2 -0.01348186       2        L FALSE
#> 277 -0.5920428959     2 -0.01348186       2        L FALSE
#> 278 -0.2072163691     2 -0.01348186       2        L FALSE
#> 279 -0.6839189198     2 -0.01348186       2        L FALSE
#> 280 -0.3443044744     2 -0.01348186       2        L FALSE
#> 281 -0.2405121721     2 -0.01348186       2        L FALSE
#> 282 -0.0544228875     2 -0.01348186       2        L FALSE
#> 283 -0.4150556072     2 -0.01348186       2        L FALSE
#> 284 -0.6059897886     2 -0.01348186       2        L FALSE
#> 285 -0.1175511975     2 -0.01348186       2        L FALSE
#> 286 -0.3967657227     2 -0.01348186       2        L FALSE
#> 287 -0.3015367007     2 -0.01348186       2        L FALSE
#> 288  0.0523087195     2 -0.01348186       2        L  TRUE
#> 289 -0.1785614560     2 -0.01348186       2        L FALSE
#> 290 -0.0544228875     2 -0.01348186       2        L FALSE
#> 291 -0.2072163691     2 -0.01348186       2        L FALSE
#> 292 -0.4289988188     2 -0.01348186       2        L FALSE
#> 293 -0.7157972196     2 -0.01348186       2        L FALSE
#> 294 -0.3443044744     2 -0.01348186       2        L FALSE
#> 295 -0.1545184385     2 -0.01348186       2        L FALSE
#> 296 -0.3443044744     2 -0.01348186       2        L FALSE
#> 297 -0.4000782290     2 -0.01348186       2        L FALSE
#> 298 -0.2405121721     2 -0.01348186       2        L FALSE
#> 299 -0.1175511975     2 -0.01348186       2        L FALSE
#> 300 -0.1545184385     2 -0.01348186       2        L FALSE
#> 301  0.3694525754     3 -0.01348186       2        R  TRUE
#> 302  0.3591128602     3 -0.01348186       2        R  TRUE
#> 303  0.2441838126     3 -0.01348186       2        R  TRUE
#> 304  0.3574329072     3 -0.01348186       2        R  TRUE
#> 305  0.2370164032     3 -0.01348186       2        R  TRUE
#> 306  0.2256498836     3 -0.01348186       2        R  TRUE
#> 307  0.2028593096     3 -0.01348186       2        R  TRUE
#> 308  0.5699167163     3 -0.01348186       2        R  TRUE
#> 309  0.2729572938     3 -0.01348186       2        R  TRUE
#> 310  0.2931886730     3 -0.01348186       2        R  TRUE
#> 311  0.2961606531     3 -0.01348186       2        R  TRUE
#> 312  0.3694525754     3 -0.01348186       2        R  TRUE
#> 313  0.4412880474     3 -0.01348186       2        R  TRUE
#> 314  0.4529524651     3 -0.01348186       2        R  TRUE
#> 315  0.3648623122     3 -0.01348186       2        R  TRUE
#> 316  0.3574329072     3 -0.01348186       2        R  TRUE
#> 317  0.0382648909     3 -0.01348186       2        R  TRUE
#> 318  0.3661617014     3 -0.01348186       2        R  TRUE
#> 319  0.2166300994     3 -0.01348186       2        R  TRUE
#> 320  0.2028593096     3 -0.01348186       2        R  TRUE
#> 321  0.2002605313     3 -0.01348186       2        R  TRUE
#> 322  0.5699167163     3 -0.01348186       2        R  TRUE
#> 323  0.4529524651     3 -0.01348186       2        R  TRUE
#> 324  0.3779410062     3 -0.01348186       2        R  TRUE
#> 325  0.4105583473     3 -0.01348186       2        R  TRUE
#> 326  0.7178832523     3 -0.01348186       2        R  TRUE
#> 327  0.7133752008     3 -0.01348186       2        R  TRUE
#> 328  0.2370164032     3 -0.01348186       2        R  TRUE
#> 329  0.3694525754     3 -0.01348186       2        R  TRUE
#> 330  0.1711032592     3 -0.01348186       2        R  TRUE
#> 331  0.4244050771     3 -0.01348186       2        R  TRUE
#> 332  0.3574329072     3 -0.01348186       2        R  TRUE
#> 333  0.4062158095     3 -0.01348186       2        R  TRUE
#> 334  0.4382159044     3 -0.01348186       2        R  TRUE
#> 335  0.3694525754     3 -0.01348186       2        R  TRUE
#> 336  0.3574329072     3 -0.01348186       2        R  TRUE
#> 337  0.4466500274     3 -0.01348186       2        R  TRUE
#> 338  0.2729572938     3 -0.01348186       2        R  TRUE
#> 339  0.3042252554     3 -0.01348186       2        R  TRUE
#> 340  0.2441838126     3 -0.01348186       2        R  TRUE
#> 341  0.0432136316     3 -0.01348186       2        R  TRUE
#> 342  0.2931886730     3 -0.01348186       2        R  TRUE
#> 343  0.2002605313     3 -0.01348186       2        R  TRUE
#> 344  0.4382159044     3 -0.01348186       2        R  TRUE
#> 345  0.0382648909     3 -0.01348186       2        R  TRUE
#> 346  0.4105583473     3 -0.01348186       2        R  TRUE
#> 347  0.2931886730     3 -0.01348186       2        R  TRUE
#> 348  0.0844485608     3 -0.01348186       2        R  TRUE
#> 349  0.7178832523     3 -0.01348186       2        R  TRUE
#> 350  0.2823429177     3 -0.01348186       2        R  TRUE
#> 351 -0.0760430765     1 -0.01665642       5        L FALSE
#> 352 -0.0585741182     1 -0.01665642       5        L FALSE
#> 353 -0.0816491639     1 -0.01665642       5        L FALSE
#> 354 -0.0760430765     1 -0.01665642       5        L FALSE
#> 355 -0.1038609232     1 -0.01665642       5        L FALSE
#> 356 -0.0741787031     1 -0.01665642       5        L FALSE
#> 357 -0.0169864945     1 -0.01665642       5        L FALSE
#> 358 -0.0868804330     1 -0.01665642       5        L FALSE
#> 359 -0.0837321729     1 -0.01665642       5        L FALSE
#> 360 -0.1130076674     1 -0.01665642       5        L FALSE
#> 361 -0.1753676097     1 -0.01665642       5        L FALSE
#> 362 -0.0922935718     1 -0.01665642       5        L FALSE
#> 363 -0.0837321729     1 -0.01665642       5        L FALSE
#> 364 -0.1744270876     1 -0.01665642       5        L FALSE
#> 365 -0.0808990737     1 -0.01665642       5        L FALSE
#> 366 -0.0964191201     1 -0.01665642       5        L FALSE
#> 367 -0.1350872179     1 -0.01665642       5        L FALSE
#> 368 -0.1279485889     1 -0.01665642       5        L FALSE
#> 369 -0.0918949547     1 -0.01665642       5        L FALSE
#> 370 -0.0918949547     1 -0.01665642       5        L FALSE
#> 371 -0.0169864945     1 -0.01665642       5        L FALSE
#> 372 -0.0808990737     1 -0.01665642       5        L FALSE
#> 373 -0.0667225712     1 -0.01665642       5        L FALSE
#> 374 -0.0741787031     1 -0.01665642       5        L FALSE
#> 375 -0.1744270876     1 -0.01665642       5        L FALSE
#> 376 -0.1737257503     1 -0.01665642       5        L FALSE
#> 377 -0.1409038324     1 -0.01665642       5        L FALSE
#> 378 -0.1279485889     1 -0.01665642       5        L FALSE
#> 379 -0.0741787031     1 -0.01665642       5        L FALSE
#> 380 -0.0169864945     1 -0.01665642       5        L FALSE
#> 381 -0.0837961623     1 -0.01665642       5        L FALSE
#> 382 -0.1352027560     1 -0.01665642       5        L FALSE
#> 383 -0.1130076674     1 -0.01665642       5        L FALSE
#> 384 -0.1092941572     1 -0.01665642       5        L FALSE
#> 385 -0.1544219839     1 -0.01665642       5        L FALSE
#> 386 -0.0947777145     1 -0.01665642       5        L FALSE
#> 387 -0.0420352314     1 -0.01665642       5        L FALSE
#> 388 -0.0816491639     1 -0.01665642       5        L FALSE
#> 389 -0.0816491639     1 -0.01665642       5        L FALSE
#> 390 -0.0760430765     1 -0.01665642       5        L FALSE
#> 391 -0.1279485889     1 -0.01665642       5        L FALSE
#> 392 -0.0918949547     1 -0.01665642       5        L FALSE
#> 393 -0.0947777145     1 -0.01665642       5        L FALSE
#> 394 -0.0738267761     1 -0.01665642       5        L FALSE
#> 395 -0.1744270876     1 -0.01665642       5        L FALSE
#> 396 -0.1158675361     1 -0.01665642       5        L FALSE
#> 397 -0.1872873421     1 -0.01665642       5        L FALSE
#> 398 -0.1360335064     1 -0.01665642       5        L FALSE
#> 399 -0.1038609232     1 -0.01665642       5        L FALSE
#> 400 -0.2121999917     1 -0.01665642       5        L FALSE
#> 401  0.0837460259     3 -0.01665642       5        R  TRUE
#> 402  0.0311190810     3 -0.01665642       5        R  TRUE
#> 403 -0.0214299997     3 -0.01665642       5        R FALSE
#> 404  0.0399975290     3 -0.01665642       5        R  TRUE
#> 405  0.0934499118     3 -0.01665642       5        R  TRUE
#> 406  0.1615376998     3 -0.01665642       5        R  TRUE
#> 407  0.1222842568     3 -0.01665642       5        R  TRUE
#> 408  0.1331965070     3 -0.01665642       5        R  TRUE
#> 409 -0.0017319484     3 -0.01665642       5        R  TRUE
#> 410  0.0517816279     3 -0.01665642       5        R  TRUE
#> 411  0.0542375669     3 -0.01665642       5        R  TRUE
#> 412  0.0837460259     3 -0.01665642       5        R  TRUE
#> 413  0.0905012744     3 -0.01665642       5        R  TRUE
#> 414  0.1945701449     3 -0.01665642       5        R  TRUE
#> 415  0.0599243145     3 -0.01665642       5        R  TRUE
#> 416  0.0399975290     3 -0.01665642       5        R  TRUE
#> 417  0.0305046242     3 -0.01665642       5        R  TRUE
#> 418  0.0775025907     3 -0.01665642       5        R  TRUE
#> 419  0.0774672589     3 -0.01665642       5        R  TRUE
#> 420  0.1222842568     3 -0.01665642       5        R  TRUE
#> 421  0.0781829250     3 -0.01665642       5        R  TRUE
#> 422  0.1331965070     3 -0.01665642       5        R  TRUE
#> 423  0.1945701449     3 -0.01665642       5        R  TRUE
#> 424  0.1334682277     3 -0.01665642       5        R  TRUE
#> 425  0.0456040702     3 -0.01665642       5        R  TRUE
#> 426  0.0582279357     3 -0.01665642       5        R  TRUE
#> 427 -0.1074732241     3 -0.01665642       5        R FALSE
#> 428  0.0934499118     3 -0.01665642       5        R  TRUE
#> 429  0.0837460259     3 -0.01665642       5        R  TRUE
#> 430  0.0519200572     3 -0.01665642       5        R  TRUE
#> 431  0.1739362556     3 -0.01665642       5        R  TRUE
#> 432  0.0399975290     3 -0.01665642       5        R  TRUE
#> 433  0.0158091078     3 -0.01665642       5        R  TRUE
#> 434  0.0502437737     3 -0.01665642       5        R  TRUE
#> 435  0.0837460259     3 -0.01665642       5        R  TRUE
#> 436  0.0399975290     3 -0.01665642       5        R  TRUE
#> 437  0.0201719528     3 -0.01665642       5        R  TRUE
#> 438 -0.0017319484     3 -0.01665642       5        R  TRUE
#> 439  0.0689330833     3 -0.01665642       5        R  TRUE
#> 440 -0.0214299997     3 -0.01665642       5        R FALSE
#> 441  0.0806961793     3 -0.01665642       5        R  TRUE
#> 442  0.0517816279     3 -0.01665642       5        R  TRUE
#> 443  0.0781829250     3 -0.01665642       5        R  TRUE
#> 444  0.0502437737     3 -0.01665642       5        R  TRUE
#> 445  0.0305046242     3 -0.01665642       5        R  TRUE
#> 446  0.0456040702     3 -0.01665642       5        R  TRUE
#> 447  0.0517816279     3 -0.01665642       5        R  TRUE
#> 448  0.1620652761     3 -0.01665642       5        R  TRUE
#> 449  0.0582279357     3 -0.01665642       5        R  TRUE
#> 450  0.1238970047     3 -0.01665642       5        R  TRUE