Data structure with the projected and boundary by node and class.

Data structure with the projected and boundary by node and class.

Usage

node_data(ppf, tr, Rule = 1)

Arguments

ppf

is a PPforest object

tr

numerical value to identify a tree

Rule

split rule 1:mean of two group means, 2:weighted mean, 3: mean of max(left group) and min(right group), 4: weighted mean of max(left group) and min(right group)

Value

Data frame with projected data for each class and node id and the boundaries

Examples

#crab data set with all the observations used as training

pprf.crab <- PPforest(data = crab, std =TRUE, class = 'Type',
 size.tr = 1, m = 200, size.p = .5, PPmethod = 'LDA')
node_data(ppf = pprf.crab, tr = 1) 
#>        proj.data Class          cut node.id LR.class   Dir
#> 1   -0.204956308     1 -0.007729725       1        L FALSE
#> 2   -0.305763986     1 -0.007729725       1        L FALSE
#> 3   -0.514491600     1 -0.007729725       1        L FALSE
#> 4   -0.556951659     1 -0.007729725       1        L FALSE
#> 5   -0.272445371     1 -0.007729725       1        L FALSE
#> 6   -0.181684995     1 -0.007729725       1        L FALSE
#> 7    0.015630358     1 -0.007729725       1        L  TRUE
#> 8   -0.734321046     1 -0.007729725       1        L FALSE
#> 9   -0.519724930     1 -0.007729725       1        L FALSE
#> 10  -0.331551020     1 -0.007729725       1        L FALSE
#> 11  -0.154789372     1 -0.007729725       1        L FALSE
#> 12  -0.204956308     1 -0.007729725       1        L FALSE
#> 13  -0.170826414     1 -0.007729725       1        L FALSE
#> 14   0.040348933     1 -0.007729725       1        L  TRUE
#> 15  -0.362813825     1 -0.007729725       1        L FALSE
#> 16  -0.368276898     1 -0.007729725       1        L FALSE
#> 17  -0.366424207     1 -0.007729725       1        L FALSE
#> 18  -0.439917326     1 -0.007729725       1        L FALSE
#> 19  -0.115913070     1 -0.007729725       1        L FALSE
#> 20  -0.440268696     1 -0.007729725       1        L FALSE
#> 21  -0.734321046     1 -0.007729725       1        L FALSE
#> 22  -0.237233511     1 -0.007729725       1        L FALSE
#> 23  -0.323302350     1 -0.007729725       1        L FALSE
#> 24  -0.305763986     1 -0.007729725       1        L FALSE
#> 25  -0.128650968     1 -0.007729725       1        L FALSE
#> 26  -0.262871874     1 -0.007729725       1        L FALSE
#> 27  -0.210581560     1 -0.007729725       1        L FALSE
#> 28   0.177206375     1 -0.007729725       1        L  TRUE
#> 29   0.015630358     1 -0.007729725       1        L  TRUE
#> 30  -0.246631712     1 -0.007729725       1        L FALSE
#> 31  -0.237233511     1 -0.007729725       1        L FALSE
#> 32  -0.440268696     1 -0.007729725       1        L FALSE
#> 33   0.040348933     1 -0.007729725       1        L  TRUE
#> 34  -0.210581560     1 -0.007729725       1        L FALSE
#> 35  -0.210581560     1 -0.007729725       1        L FALSE
#> 36  -0.145486171     1 -0.007729725       1        L FALSE
#> 37   0.044784102     1 -0.007729725       1        L  TRUE
#> 38  -0.474331024     1 -0.007729725       1        L FALSE
#> 39  -0.204956308     1 -0.007729725       1        L FALSE
#> 40  -0.246631712     1 -0.007729725       1        L FALSE
#> 41  -0.246631712     1 -0.007729725       1        L FALSE
#> 42  -0.366424207     1 -0.007729725       1        L FALSE
#> 43  -0.070573224     1 -0.007729725       1        L FALSE
#> 44  -0.439917326     1 -0.007729725       1        L FALSE
#> 45  -0.474331024     1 -0.007729725       1        L FALSE
#> 46  -0.246631712     1 -0.007729725       1        L FALSE
#> 47  -0.411899606     1 -0.007729725       1        L FALSE
#> 48  -0.323640213     1 -0.007729725       1        L FALSE
#> 49   0.044784102     1 -0.007729725       1        L  TRUE
#> 50  -0.428774941     1 -0.007729725       1        L FALSE
#> 51  -0.117386958     2 -0.007729725       1        R FALSE
#> 52   0.035291677     2 -0.007729725       1        R  TRUE
#> 53   0.274579744     2 -0.007729725       1        R  TRUE
#> 54  -0.117386958     2 -0.007729725       1        R FALSE
#> 55  -0.122768959     2 -0.007729725       1        R FALSE
#> 56   0.304490320     2 -0.007729725       1        R  TRUE
#> 57   0.554595994     2 -0.007729725       1        R  TRUE
#> 58   0.082970749     2 -0.007729725       1        R  TRUE
#> 59   0.206197941     2 -0.007729725       1        R  TRUE
#> 60   0.188443340     2 -0.007729725       1        R  TRUE
#> 61   0.274579744     2 -0.007729725       1        R  TRUE
#> 62   0.167997754     2 -0.007729725       1        R  TRUE
#> 63   0.098020827     2 -0.007729725       1        R  TRUE
#> 64   0.310602073     2 -0.007729725       1        R  TRUE
#> 65   0.231700750     2 -0.007729725       1        R  TRUE
#> 66   0.554595994     2 -0.007729725       1        R  TRUE
#> 67   0.160114382     2 -0.007729725       1        R  TRUE
#> 68   0.082970749     2 -0.007729725       1        R  TRUE
#> 69   0.132380466     2 -0.007729725       1        R  TRUE
#> 70   0.251118430     2 -0.007729725       1        R  TRUE
#> 71  -0.083852105     2 -0.007729725       1        R FALSE
#> 72   0.206197941     2 -0.007729725       1        R  TRUE
#> 73  -0.036321704     2 -0.007729725       1        R FALSE
#> 74   0.364150036     2 -0.007729725       1        R  TRUE
#> 75   0.209984008     2 -0.007729725       1        R  TRUE
#> 76   0.166118438     2 -0.007729725       1        R  TRUE
#> 77  -0.044516315     2 -0.007729725       1        R FALSE
#> 78  -0.048302382     2 -0.007729725       1        R FALSE
#> 79   0.035291677     2 -0.007729725       1        R  TRUE
#> 80   0.554595994     2 -0.007729725       1        R  TRUE
#> 81  -0.083852105     2 -0.007729725       1        R FALSE
#> 82   0.251118430     2 -0.007729725       1        R  TRUE
#> 83   0.310602073     2 -0.007729725       1        R  TRUE
#> 84   0.368112210     2 -0.007729725       1        R  TRUE
#> 85  -0.036321704     2 -0.007729725       1        R FALSE
#> 86   0.209984008     2 -0.007729725       1        R  TRUE
#> 87   0.099264970     2 -0.007729725       1        R  TRUE
#> 88   0.188443340     2 -0.007729725       1        R  TRUE
#> 89   0.035291677     2 -0.007729725       1        R  TRUE
#> 90   0.099535267     2 -0.007729725       1        R  TRUE
#> 91   0.132380466     2 -0.007729725       1        R  TRUE
#> 92   0.170445487     2 -0.007729725       1        R  TRUE
#> 93   0.206197941     2 -0.007729725       1        R  TRUE
#> 94  -0.083852105     2 -0.007729725       1        R FALSE
#> 95   0.153596778     2 -0.007729725       1        R  TRUE
#> 96   0.099264970     2 -0.007729725       1        R  TRUE
#> 97   0.058198114     2 -0.007729725       1        R  TRUE
#> 98   0.342311669     2 -0.007729725       1        R  TRUE
#> 99  -0.117386958     2 -0.007729725       1        R FALSE
#> 100  0.206197941     2 -0.007729725       1        R  TRUE
#> 101  0.004542032     3 -0.007729725       1        L  TRUE
#> 102 -0.321719956     3 -0.007729725       1        L FALSE
#> 103 -0.095277905     3 -0.007729725       1        L FALSE
#> 104 -0.427787588     3 -0.007729725       1        L FALSE
#> 105 -0.364287712     3 -0.007729725       1        L FALSE
#> 106  0.153475152     3 -0.007729725       1        L  TRUE
#> 107 -0.436008824     3 -0.007729725       1        L FALSE
#> 108 -0.178912517     3 -0.007729725       1        L FALSE
#> 109 -0.008087359     3 -0.007729725       1        L FALSE
#> 110 -0.700718627     3 -0.007729725       1        L FALSE
#> 111 -0.427787588     3 -0.007729725       1        L FALSE
#> 112 -0.275676949     3 -0.007729725       1        L FALSE
#> 113  0.004542032     3 -0.007729725       1        L  TRUE
#> 114 -0.161239021     3 -0.007729725       1        L FALSE
#> 115 -0.436008824     3 -0.007729725       1        L FALSE
#> 116 -0.711374089     3 -0.007729725       1        L FALSE
#> 117 -0.406625335     3 -0.007729725       1        L FALSE
#> 118 -0.311064494     3 -0.007729725       1        L FALSE
#> 119 -0.394996837     3 -0.007729725       1        L FALSE
#> 120 -0.166932228     3 -0.007729725       1        L FALSE
#> 121 -0.095277905     3 -0.007729725       1        L FALSE
#> 122 -0.865174432     3 -0.007729725       1        L FALSE
#> 123 -0.711374089     3 -0.007729725       1        L FALSE
#> 124 -0.427787588     3 -0.007729725       1        L FALSE
#> 125 -0.388018935     3 -0.007729725       1        L FALSE
#> 126 -0.368655314     3 -0.007729725       1        L FALSE
#> 127 -0.388018935     3 -0.007729725       1        L FALSE
#> 128 -0.178912517     3 -0.007729725       1        L FALSE
#> 129 -0.311064494     3 -0.007729725       1        L FALSE
#> 130 -0.422987122     3 -0.007729725       1        L FALSE
#> 131 -0.324221748     3 -0.007729725       1        L FALSE
#> 132 -0.436008824     3 -0.007729725       1        L FALSE
#> 133 -0.364287712     3 -0.007729725       1        L FALSE
#> 134 -0.095277905     3 -0.007729725       1        L FALSE
#> 135 -0.295838731     3 -0.007729725       1        L FALSE
#> 136  0.140615629     3 -0.007729725       1        L  TRUE
#> 137 -0.178912517     3 -0.007729725       1        L FALSE
#> 138 -0.325587517     3 -0.007729725       1        L FALSE
#> 139 -0.865174432     3 -0.007729725       1        L FALSE
#> 140 -0.295838731     3 -0.007729725       1        L FALSE
#> 141 -0.355930535     3 -0.007729725       1        L FALSE
#> 142 -0.293945487     3 -0.007729725       1        L FALSE
#> 143 -0.364287712     3 -0.007729725       1        L FALSE
#> 144 -0.052615925     3 -0.007729725       1        L FALSE
#> 145 -0.008087359     3 -0.007729725       1        L FALSE
#> 146 -0.324221748     3 -0.007729725       1        L FALSE
#> 147 -0.115561312     3 -0.007729725       1        L FALSE
#> 148 -0.369534160     3 -0.007729725       1        L FALSE
#> 149 -0.422987122     3 -0.007729725       1        L FALSE
#> 150  0.004542032     3 -0.007729725       1        L  TRUE
#> 151  0.373034301     4 -0.007729725       1        R  TRUE
#> 152  0.519154391     4 -0.007729725       1        R  TRUE
#> 153  0.329344027     4 -0.007729725       1        R  TRUE
#> 154  0.485930776     4 -0.007729725       1        R  TRUE
#> 155  0.121819616     4 -0.007729725       1        R  TRUE
#> 156  0.209700205     4 -0.007729725       1        R  TRUE
#> 157  0.477911850     4 -0.007729725       1        R  TRUE
#> 158  0.382567246     4 -0.007729725       1        R  TRUE
#> 159  0.442199949     4 -0.007729725       1        R  TRUE
#> 160  0.448595928     4 -0.007729725       1        R  TRUE
#> 161  0.275945124     4 -0.007729725       1        R  TRUE
#> 162  0.381039299     4 -0.007729725       1        R  TRUE
#> 163  0.275945124     4 -0.007729725       1        R  TRUE
#> 164  0.270657735     4 -0.007729725       1        R  TRUE
#> 165  0.545725658     4 -0.007729725       1        R  TRUE
#> 166  0.485930776     4 -0.007729725       1        R  TRUE
#> 167  0.519154391     4 -0.007729725       1        R  TRUE
#> 168  0.169012153     4 -0.007729725       1        R  TRUE
#> 169  0.696821898     4 -0.007729725       1        R  TRUE
#> 170  0.329344027     4 -0.007729725       1        R  TRUE
#> 171  0.286451915     4 -0.007729725       1        R  TRUE
#> 172  0.365732043     4 -0.007729725       1        R  TRUE
#> 173  0.121819616     4 -0.007729725       1        R  TRUE
#> 174  0.477911850     4 -0.007729725       1        R  TRUE
#> 175  0.448595928     4 -0.007729725       1        R  TRUE
#> 176  0.656566321     4 -0.007729725       1        R  TRUE
#> 177  0.571241585     4 -0.007729725       1        R  TRUE
#> 178  0.559355519     4 -0.007729725       1        R  TRUE
#> 179  0.435898583     4 -0.007729725       1        R  TRUE
#> 180  0.571241585     4 -0.007729725       1        R  TRUE
#> 181  0.169012153     4 -0.007729725       1        R  TRUE
#> 182  0.381039299     4 -0.007729725       1        R  TRUE
#> 183  0.519154391     4 -0.007729725       1        R  TRUE
#> 184  0.448595928     4 -0.007729725       1        R  TRUE
#> 185  0.355509444     4 -0.007729725       1        R  TRUE
#> 186  0.386164121     4 -0.007729725       1        R  TRUE
#> 187  0.355509444     4 -0.007729725       1        R  TRUE
#> 188  0.365732043     4 -0.007729725       1        R  TRUE
#> 189  0.209700205     4 -0.007729725       1        R  TRUE
#> 190  0.410950263     4 -0.007729725       1        R  TRUE
#> 191  0.386880789     4 -0.007729725       1        R  TRUE
#> 192  0.460076176     4 -0.007729725       1        R  TRUE
#> 193  0.270657735     4 -0.007729725       1        R  TRUE
#> 194  0.431206236     4 -0.007729725       1        R  TRUE
#> 195  0.390991601     4 -0.007729725       1        R  TRUE
#> 196  0.460076176     4 -0.007729725       1        R  TRUE
#> 197  0.280975334     4 -0.007729725       1        R  TRUE
#> 198  0.552202320     4 -0.007729725       1        R  TRUE
#> 199  0.365732043     4 -0.007729725       1        R  TRUE
#> 200  0.191228936     4 -0.007729725       1        R  TRUE
#> 201 -0.264150888     1 -0.151446963       2        L FALSE
#> 202 -0.516981880     1 -0.151446963       2        L FALSE
#> 203 -0.265430491     1 -0.151446963       2        L FALSE
#> 204 -0.381514308     1 -0.151446963       2        L FALSE
#> 205 -0.197663314     1 -0.151446963       2        L FALSE
#> 206 -0.244770465     1 -0.151446963       2        L FALSE
#> 207 -0.493272589     1 -0.151446963       2        L FALSE
#> 208 -0.288245811     1 -0.151446963       2        L FALSE
#> 209 -0.365063777     1 -0.151446963       2        L FALSE
#> 210 -0.410331153     1 -0.151446963       2        L FALSE
#> 211 -0.452549266     1 -0.151446963       2        L FALSE
#> 212 -0.264150888     1 -0.151446963       2        L FALSE
#> 213 -0.272325986     1 -0.151446963       2        L FALSE
#> 214 -0.398428007     1 -0.151446963       2        L FALSE
#> 215 -0.377378257     1 -0.151446963       2        L FALSE
#> 216 -0.476859375     1 -0.151446963       2        L FALSE
#> 217 -0.326142909     1 -0.151446963       2        L FALSE
#> 218 -0.409425998     1 -0.151446963       2        L FALSE
#> 219 -0.486937265     1 -0.151446963       2        L FALSE
#> 220 -0.560234946     1 -0.151446963       2        L FALSE
#> 221 -0.288245811     1 -0.151446963       2        L FALSE
#> 222 -0.105259324     1 -0.151446963       2        L  TRUE
#> 223 -0.336666118     1 -0.151446963       2        L FALSE
#> 224 -0.516981880     1 -0.151446963       2        L FALSE
#> 225 -0.408391045     1 -0.151446963       2        L FALSE
#> 226 -0.527352921     1 -0.151446963       2        L FALSE
#> 227 -0.136962081     1 -0.151446963       2        L  TRUE
#> 228 -0.582393851     1 -0.151446963       2        L FALSE
#> 229 -0.493272589     1 -0.151446963       2        L FALSE
#> 230 -0.506021205     1 -0.151446963       2        L FALSE
#> 231 -0.105259324     1 -0.151446963       2        L  TRUE
#> 232 -0.560234946     1 -0.151446963       2        L FALSE
#> 233 -0.398428007     1 -0.151446963       2        L FALSE
#> 234 -0.136962081     1 -0.151446963       2        L  TRUE
#> 235 -0.136962081     1 -0.151446963       2        L  TRUE
#> 236 -0.404455664     1 -0.151446963       2        L FALSE
#> 237 -0.475082619     1 -0.151446963       2        L FALSE
#> 238 -0.325542091     1 -0.151446963       2        L FALSE
#> 239 -0.264150888     1 -0.151446963       2        L FALSE
#> 240 -0.506021205     1 -0.151446963       2        L FALSE
#> 241 -0.506021205     1 -0.151446963       2        L FALSE
#> 242 -0.326142909     1 -0.151446963       2        L FALSE
#> 243 -0.371858937     1 -0.151446963       2        L FALSE
#> 244 -0.409425998     1 -0.151446963       2        L FALSE
#> 245 -0.325542091     1 -0.151446963       2        L FALSE
#> 246 -0.506021205     1 -0.151446963       2        L FALSE
#> 247 -0.416258476     1 -0.151446963       2        L FALSE
#> 248 -0.185400667     1 -0.151446963       2        L FALSE
#> 249 -0.475082619     1 -0.151446963       2        L FALSE
#> 250 -0.512690329     1 -0.151446963       2        L FALSE
#> 251 -0.465564900     3 -0.151446963       2        R FALSE
#> 252  0.191833555     3 -0.151446963       2        R  TRUE
#> 253  0.641226839     3 -0.151446963       2        R  TRUE
#> 254  0.041076441     3 -0.151446963       2        R  TRUE
#> 255 -0.251135253     3 -0.151446963       2        R FALSE
#> 256 -0.123471663     3 -0.151446963       2        R  TRUE
#> 257  0.093561928     3 -0.151446963       2        R  TRUE
#> 258  0.233195015     3 -0.151446963       2        R  TRUE
#> 259  0.267723997     3 -0.151446963       2        R  TRUE
#> 260 -0.044636055     3 -0.151446963       2        R  TRUE
#> 261  0.041076441     3 -0.151446963       2        R  TRUE
#> 262  0.130531503     3 -0.151446963       2        R  TRUE
#> 263 -0.465564900     3 -0.151446963       2        R FALSE
#> 264  0.505914767     3 -0.151446963       2        R  TRUE
#> 265  0.093561928     3 -0.151446963       2        R  TRUE
#> 266  0.184474462     3 -0.151446963       2        R  TRUE
#> 267  0.111599730     3 -0.151446963       2        R  TRUE
#> 268 -0.037276962     3 -0.151446963       2        R  TRUE
#> 269 -0.193676100     3 -0.151446963       2        R FALSE
#> 270 -0.399270143     3 -0.151446963       2        R FALSE
#> 271  0.641226839     3 -0.151446963       2        R  TRUE
#> 272  0.045490696     3 -0.151446963       2        R  TRUE
#> 273  0.184474462     3 -0.151446963       2        R  TRUE
#> 274  0.041076441     3 -0.151446963       2        R  TRUE
#> 275  0.207786934     3 -0.151446963       2        R  TRUE
#> 276  0.052037116     3 -0.151446963       2        R  TRUE
#> 277  0.207786934     3 -0.151446963       2        R  TRUE
#> 278  0.233195015     3 -0.151446963       2        R  TRUE
#> 279 -0.037276962     3 -0.151446963       2        R  TRUE
#> 280 -0.060693101     3 -0.151446963       2        R  TRUE
#> 281  0.141940829     3 -0.151446963       2        R  TRUE
#> 282  0.093561928     3 -0.151446963       2        R  TRUE
#> 283 -0.251135253     3 -0.151446963       2        R FALSE
#> 284  0.641226839     3 -0.151446963       2        R  TRUE
#> 285  0.109993422     3 -0.151446963       2        R  TRUE
#> 286 -0.195886559     3 -0.151446963       2        R FALSE
#> 287  0.233195015     3 -0.151446963       2        R  TRUE
#> 288  0.117070549     3 -0.151446963       2        R  TRUE
#> 289  0.045490696     3 -0.151446963       2        R  TRUE
#> 290  0.109993422     3 -0.151446963       2        R  TRUE
#> 291  0.477272459     3 -0.151446963       2        R  TRUE
#> 292 -0.116921152     3 -0.151446963       2        R  TRUE
#> 293 -0.251135253     3 -0.151446963       2        R FALSE
#> 294 -0.174847994     3 -0.151446963       2        R FALSE
#> 295  0.267723997     3 -0.151446963       2        R  TRUE
#> 296  0.141940829     3 -0.151446963       2        R  TRUE
#> 297  0.469727643     3 -0.151446963       2        R  TRUE
#> 298  0.153013023     3 -0.151446963       2        R  TRUE
#> 299 -0.060693101     3 -0.151446963       2        R  TRUE
#> 300 -0.465564900     3 -0.151446963       2        R FALSE
#> 301 -0.070072422     2 -0.007150655       3        L FALSE
#> 302 -0.109475869     2 -0.007150655       3        L FALSE
#> 303 -0.072571489     2 -0.007150655       3        L FALSE
#> 304 -0.070072422     2 -0.007150655       3        L FALSE
#> 305 -0.118797645     2 -0.007150655       3        L FALSE
#> 306 -0.107535232     2 -0.007150655       3        L FALSE
#> 307 -0.101638970     2 -0.007150655       3        L FALSE
#> 308 -0.043936282     2 -0.007150655       3        L FALSE
#> 309 -0.068552660     2 -0.007150655       3        L FALSE
#> 310 -0.090473518     2 -0.007150655       3        L FALSE
#> 311 -0.072571489     2 -0.007150655       3        L FALSE
#> 312  0.002674428     2 -0.007150655       3        L  TRUE
#> 313 -0.031906415     2 -0.007150655       3        L FALSE
#> 314 -0.134141664     2 -0.007150655       3        L FALSE
#> 315 -0.072657666     2 -0.007150655       3        L FALSE
#> 316 -0.101638970     2 -0.007150655       3        L FALSE
#> 317 -0.153960668     2 -0.007150655       3        L FALSE
#> 318 -0.043936282     2 -0.007150655       3        L FALSE
#> 319 -0.019296066     2 -0.007150655       3        L FALSE
#> 320 -0.076750057     2 -0.007150655       3        L FALSE
#> 321 -0.075574605     2 -0.007150655       3        L FALSE
#> 322 -0.068552660     2 -0.007150655       3        L FALSE
#> 323 -0.050006200     2 -0.007150655       3        L FALSE
#> 324 -0.173557945     2 -0.007150655       3        L FALSE
#> 325 -0.071692849     2 -0.007150655       3        L FALSE
#> 326 -0.114187111     2 -0.007150655       3        L FALSE
#> 327 -0.106013596     2 -0.007150655       3        L FALSE
#> 328 -0.018961192     2 -0.007150655       3        L FALSE
#> 329 -0.109475869     2 -0.007150655       3        L FALSE
#> 330 -0.101638970     2 -0.007150655       3        L FALSE
#> 331 -0.075574605     2 -0.007150655       3        L FALSE
#> 332 -0.076750057     2 -0.007150655       3        L FALSE
#> 333 -0.134141664     2 -0.007150655       3        L FALSE
#> 334 -0.133165713     2 -0.007150655       3        L FALSE
#> 335 -0.050006200     2 -0.007150655       3        L FALSE
#> 336 -0.071692849     2 -0.007150655       3        L FALSE
#> 337 -0.125858179     2 -0.007150655       3        L FALSE
#> 338 -0.090473518     2 -0.007150655       3        L FALSE
#> 339 -0.109475869     2 -0.007150655       3        L FALSE
#> 340 -0.049920199     2 -0.007150655       3        L FALSE
#> 341 -0.019296066     2 -0.007150655       3        L FALSE
#> 342 -0.040610994     2 -0.007150655       3        L FALSE
#> 343 -0.068552660     2 -0.007150655       3        L FALSE
#> 344 -0.075574605     2 -0.007150655       3        L FALSE
#> 345 -0.067106021     2 -0.007150655       3        L FALSE
#> 346 -0.125858179     2 -0.007150655       3        L FALSE
#> 347 -0.032710233     2 -0.007150655       3        L FALSE
#> 348 -0.071211158     2 -0.007150655       3        L FALSE
#> 349 -0.070072422     2 -0.007150655       3        L FALSE
#> 350 -0.068552660     2 -0.007150655       3        L FALSE
#> 351  0.022394465     4 -0.007150655       3        R  TRUE
#> 352  0.067670293     4 -0.007150655       3        R  TRUE
#> 353  0.065877645     4 -0.007150655       3        R  TRUE
#> 354  0.106491691     4 -0.007150655       3        R  TRUE
#> 355  0.024966963     4 -0.007150655       3        R  TRUE
#> 356  0.051264101     4 -0.007150655       3        R  TRUE
#> 357  0.049594433     4 -0.007150655       3        R  TRUE
#> 358  0.035920301     4 -0.007150655       3        R  TRUE
#> 359  0.061031072     4 -0.007150655       3        R  TRUE
#> 360  0.095117435     4 -0.007150655       3        R  TRUE
#> 361  0.058694438     4 -0.007150655       3        R  TRUE
#> 362  0.076757685     4 -0.007150655       3        R  TRUE
#> 363  0.058694438     4 -0.007150655       3        R  TRUE
#> 364  0.048371131     4 -0.007150655       3        R  TRUE
#> 365  0.082122663     4 -0.007150655       3        R  TRUE
#> 366  0.106491691     4 -0.007150655       3        R  TRUE
#> 367  0.067670293     4 -0.007150655       3        R  TRUE
#> 368  0.010822364     4 -0.007150655       3        R  TRUE
#> 369  0.041631464     4 -0.007150655       3        R  TRUE
#> 370  0.065877645     4 -0.007150655       3        R  TRUE
#> 371  0.049211621     4 -0.007150655       3        R  TRUE
#> 372  0.080342849     4 -0.007150655       3        R  TRUE
#> 373  0.024966963     4 -0.007150655       3        R  TRUE
#> 374  0.049594433     4 -0.007150655       3        R  TRUE
#> 375  0.095117435     4 -0.007150655       3        R  TRUE
#> 376  0.128844954     4 -0.007150655       3        R  TRUE
#> 377  0.140677064     4 -0.007150655       3        R  TRUE
#> 378  0.041904043     4 -0.007150655       3        R  TRUE
#> 379  0.048556186     4 -0.007150655       3        R  TRUE
#> 380  0.140677064     4 -0.007150655       3        R  TRUE
#> 381  0.010822364     4 -0.007150655       3        R  TRUE
#> 382  0.076757685     4 -0.007150655       3        R  TRUE
#> 383  0.067670293     4 -0.007150655       3        R  TRUE
#> 384  0.095117435     4 -0.007150655       3        R  TRUE
#> 385  0.109904810     4 -0.007150655       3        R  TRUE
#> 386  0.009622987     4 -0.007150655       3        R  TRUE
#> 387  0.109904810     4 -0.007150655       3        R  TRUE
#> 388  0.080342849     4 -0.007150655       3        R  TRUE
#> 389  0.051264101     4 -0.007150655       3        R  TRUE
#> 390  0.087687010     4 -0.007150655       3        R  TRUE
#> 391  0.039765386     4 -0.007150655       3        R  TRUE
#> 392  0.039554927     4 -0.007150655       3        R  TRUE
#> 393  0.048371131     4 -0.007150655       3        R  TRUE
#> 394  0.001833939     4 -0.007150655       3        R  TRUE
#> 395  0.127077886     4 -0.007150655       3        R  TRUE
#> 396  0.039554927     4 -0.007150655       3        R  TRUE
#> 397  0.035982377     4 -0.007150655       3        R  TRUE
#> 398  0.031024179     4 -0.007150655       3        R  TRUE
#> 399  0.080342849     4 -0.007150655       3        R  TRUE
#> 400  0.148553900     4 -0.007150655       3        R  TRUE