Data structure with the projected and boundary by node and class.

Data structure with the projected and boundary by node and class.

Usage

node_data(ppf, tr, Rule = 1)

Arguments

ppf

is a PPforest object

tr

numerical value to identify a tree

Rule

split rule 1:mean of two group means, 2:weighted mean, 3: mean of max(left group) and min(right group), 4: weighted mean of max(left group) and min(right group)

Value

Data frame with projected data for each class and node id and the boundaries

Examples

#crab data set with all the observations used as training

pprf.crab <- PPforest(data = crab, std = 'min-max', y = 'Type',
 size.tr = 1, m = 200, size.p = .5, PPmethod = 'LDA')
node_data(ppf = pprf.crab, tr = 1) 
#>                 proj.data        Class          cut node.id LR.class   Dir
#> 99...1       0.0022341643   BlueFemale -0.021714183       1        R  TRUE
#> 80...2      -0.0037080026   BlueFemale -0.021714183       1        R  TRUE
#> 89...3      -0.0029009777   BlueFemale -0.021714183       1        R  TRUE
#> 69...4      -0.0045210440   BlueFemale -0.021714183       1        R  TRUE
#> 89.1...5    -0.0029009777   BlueFemale -0.021714183       1        R  TRUE
#> 77...6       0.0117300948   BlueFemale -0.021714183       1        R  TRUE
#> 96...7      -0.0249024946   BlueFemale -0.021714183       1        R FALSE
#> 60...8       0.0101318554   BlueFemale -0.021714183       1        R  TRUE
#> 65...9      -0.0071010544   BlueFemale -0.021714183       1        R  TRUE
#> 55...10      0.0105444582   BlueFemale -0.021714183       1        R  TRUE
#> 61...11      0.0301789424   BlueFemale -0.021714183       1        R  TRUE
#> 99.1...12    0.0022341643   BlueFemale -0.021714183       1        R  TRUE
#> 65.1...13   -0.0071010544   BlueFemale -0.021714183       1        R  TRUE
#> 87...14      0.0134036087   BlueFemale -0.021714183       1        R  TRUE
#> 90...15     -0.0065937643   BlueFemale -0.021714183       1        R  TRUE
#> 56...16     -0.0060598192   BlueFemale -0.021714183       1        R  TRUE
#> 62...17     -0.0015381810   BlueFemale -0.021714183       1        R  TRUE
#> 64...18      0.0034652744   BlueFemale -0.021714183       1        R  TRUE
#> 56.1...19   -0.0060598192   BlueFemale -0.021714183       1        R  TRUE
#> 56.2...20   -0.0060598192   BlueFemale -0.021714183       1        R  TRUE
#> 100...21    -0.0176036349   BlueFemale -0.021714183       1        R  TRUE
#> 100.1...22  -0.0176036349   BlueFemale -0.021714183       1        R  TRUE
#> 57...23      0.0094219320   BlueFemale -0.021714183       1        R  TRUE
#> 96.1...24   -0.0249024946   BlueFemale -0.021714183       1        R FALSE
#> 79...25     -0.0019847310   BlueFemale -0.021714183       1        R  TRUE
#> 70...26     -0.0089395644   BlueFemale -0.021714183       1        R  TRUE
#> 73...27      0.0209955049   BlueFemale -0.021714183       1        R  TRUE
#> 86...28      0.0105044070   BlueFemale -0.021714183       1        R  TRUE
#> 55.1...29    0.0105444582   BlueFemale -0.021714183       1        R  TRUE
#> 67...30      0.0154575103   BlueFemale -0.021714183       1        R  TRUE
#> 85...31     -0.0071380537   BlueFemale -0.021714183       1        R  TRUE
#> 66...32      0.0081088929   BlueFemale -0.021714183       1        R  TRUE
#> 92...33      0.0141414193   BlueFemale -0.021714183       1        R  TRUE
#> 61.1...34    0.0301789424   BlueFemale -0.021714183       1        R  TRUE
#> 75...35      0.0170896970   BlueFemale -0.021714183       1        R  TRUE
#> 64.1...36    0.0034652744   BlueFemale -0.021714183       1        R  TRUE
#> 60.1...37    0.0101318554   BlueFemale -0.021714183       1        R  TRUE
#> 98...38      0.0064117761   BlueFemale -0.021714183       1        R  TRUE
#> 67.1...39    0.0154575103   BlueFemale -0.021714183       1        R  TRUE
#> 74...40      0.0068237409   BlueFemale -0.021714183       1        R  TRUE
#> 52...41      0.0192188732   BlueFemale -0.021714183       1        R  TRUE
#> 78...42      0.0030744985   BlueFemale -0.021714183       1        R  TRUE
#> 83...43      0.0116773289   BlueFemale -0.021714183       1        R  TRUE
#> 80.1...44   -0.0037080026   BlueFemale -0.021714183       1        R  TRUE
#> 67.2...45    0.0154575103   BlueFemale -0.021714183       1        R  TRUE
#> 95...46      0.0241664669   BlueFemale -0.021714183       1        R  TRUE
#> 82...47     -0.0149805649   BlueFemale -0.021714183       1        R  TRUE
#> 66.1...48    0.0081088929   BlueFemale -0.021714183       1        R  TRUE
#> 70.1...49   -0.0089395644   BlueFemale -0.021714183       1        R  TRUE
#> 59...50      0.0237405114   BlueFemale -0.021714183       1        R  TRUE
#> 44...51      0.0304180749     BlueMale -0.021714183       1        R  TRUE
#> 48...52      0.0083194566     BlueMale -0.021714183       1        R  TRUE
#> 29...53      0.0114958845     BlueMale -0.021714183       1        R  TRUE
#> 17...54      0.0344415352     BlueMale -0.021714183       1        R  TRUE
#> 50...55      0.0240767016     BlueMale -0.021714183       1        R  TRUE
#> 12...56      0.0102823608     BlueMale -0.021714183       1        R  TRUE
#> 31...57     -0.0141140759     BlueMale -0.021714183       1        R  TRUE
#> 6...58       0.0025708982     BlueMale -0.021714183       1        R  TRUE
#> 25...59      0.0156516694     BlueMale -0.021714183       1        R  TRUE
#> 5...60       0.0083837487     BlueMale -0.021714183       1        R  TRUE
#> 9...61       0.0081962441     BlueMale -0.021714183       1        R  TRUE
#> 15...62      0.0229341245     BlueMale -0.021714183       1        R  TRUE
#> 30...63     -0.0042267753     BlueMale -0.021714183       1        R  TRUE
#> 37...64      0.0258460410     BlueMale -0.021714183       1        R  TRUE
#> 9.1...65     0.0081962441     BlueMale -0.021714183       1        R  TRUE
#> 44.1...66    0.0304180749     BlueMale -0.021714183       1        R  TRUE
#> 36...67      0.0214930450     BlueMale -0.021714183       1        R  TRUE
#> 39...68      0.0060519393     BlueMale -0.021714183       1        R  TRUE
#> 8...69       0.0010418295     BlueMale -0.021714183       1        R  TRUE
#> 18...70      0.0312426424     BlueMale -0.021714183       1        R  TRUE
#> 34...71      0.0114491787     BlueMale -0.021714183       1        R  TRUE
#> 27...72      0.0172620292     BlueMale -0.021714183       1        R  TRUE
#> 18.1...73    0.0312426424     BlueMale -0.021714183       1        R  TRUE
#> 13...74      0.0242132163     BlueMale -0.021714183       1        R  TRUE
#> 3...75       0.0066951062     BlueMale -0.021714183       1        R  TRUE
#> 12.1...76    0.0102823608     BlueMale -0.021714183       1        R  TRUE
#> 45...77      0.0051672696     BlueMale -0.021714183       1        R  TRUE
#> 41...78      0.0275158649     BlueMale -0.021714183       1        R  TRUE
#> 38...79      0.0128374486     BlueMale -0.021714183       1        R  TRUE
#> 8.1...80     0.0010418295     BlueMale -0.021714183       1        R  TRUE
#> 7...81       0.0140783089     BlueMale -0.021714183       1        R  TRUE
#> 49...82     -0.0026345961     BlueMale -0.021714183       1        R  TRUE
#> 22...83      0.0088066959     BlueMale -0.021714183       1        R  TRUE
#> 24...84      0.0188505621     BlueMale -0.021714183       1        R  TRUE
#> 9.2...85     0.0081962441     BlueMale -0.021714183       1        R  TRUE
#> 30.1...86   -0.0042267753     BlueMale -0.021714183       1        R  TRUE
#> 14...87      0.0208984035     BlueMale -0.021714183       1        R  TRUE
#> 12.2...88    0.0102823608     BlueMale -0.021714183       1        R  TRUE
#> 35...89      0.0727919302     BlueMale -0.021714183       1        R  TRUE
#> 15.1...90    0.0229341245     BlueMale -0.021714183       1        R  TRUE
#> 41.1...91    0.0275158649     BlueMale -0.021714183       1        R  TRUE
#> 5.1...92     0.0083837487     BlueMale -0.021714183       1        R  TRUE
#> 42...93      0.0240075310     BlueMale -0.021714183       1        R  TRUE
#> 22.1...94    0.0088066959     BlueMale -0.021714183       1        R  TRUE
#> 38.1...95    0.0128374486     BlueMale -0.021714183       1        R  TRUE
#> 34.1...96    0.0114491787     BlueMale -0.021714183       1        R  TRUE
#> 23...97      0.0002384508     BlueMale -0.021714183       1        R  TRUE
#> 32...98      0.0183506081     BlueMale -0.021714183       1        R  TRUE
#> 1...99      -0.0150127363     BlueMale -0.021714183       1        R  TRUE
#> 11...100     0.0390821454     BlueMale -0.021714183       1        R  TRUE
#> 186...101   -0.0577828334 OrangeFemale -0.021714183       1        L FALSE
#> 161...102   -0.0537556832 OrangeFemale -0.021714183       1        L FALSE
#> 191...103   -0.0800380172 OrangeFemale -0.021714183       1        L FALSE
#> 166...104   -0.0462978876 OrangeFemale -0.021714183       1        L FALSE
#> 185...105   -0.0623797462 OrangeFemale -0.021714183       1        L FALSE
#> 197...106   -0.1060515119 OrangeFemale -0.021714183       1        L FALSE
#> 156...107   -0.0632589499 OrangeFemale -0.021714183       1        L FALSE
#> 157...108   -0.0224481107 OrangeFemale -0.021714183       1        L FALSE
#> 184...109   -0.0535585158 OrangeFemale -0.021714183       1        L FALSE
#> 172...110   -0.0690153882 OrangeFemale -0.021714183       1        L FALSE
#> 192...111   -0.1044848059 OrangeFemale -0.021714183       1        L FALSE
#> 199...112   -0.1011827952 OrangeFemale -0.021714183       1        L FALSE
#> 154...113   -0.0440309646 OrangeFemale -0.021714183       1        L FALSE
#> 173...114   -0.0587403636 OrangeFemale -0.021714183       1        L FALSE
#> 186.1...115 -0.0577828334 OrangeFemale -0.021714183       1        L FALSE
#> 155...116   -0.0437682292 OrangeFemale -0.021714183       1        L FALSE
#> 200...117   -0.0759380502 OrangeFemale -0.021714183       1        L FALSE
#> 163...118   -0.0275419692 OrangeFemale -0.021714183       1        L FALSE
#> 153...119   -0.0463482396 OrangeFemale -0.021714183       1        L FALSE
#> 185.1...120 -0.0623797462 OrangeFemale -0.021714183       1        L FALSE
#> 190...121   -0.0819578182 OrangeFemale -0.021714183       1        L FALSE
#> 168...122   -0.0404910974 OrangeFemale -0.021714183       1        L FALSE
#> 169...123   -0.0647722520 OrangeFemale -0.021714183       1        L FALSE
#> 165...124   -0.0447093547 OrangeFemale -0.021714183       1        L FALSE
#> 154.1...125 -0.0440309646 OrangeFemale -0.021714183       1        L FALSE
#> 169.1...126 -0.0647722520 OrangeFemale -0.021714183       1        L FALSE
#> 159...127   -0.0270699023 OrangeFemale -0.021714183       1        L FALSE
#> 177...128   -0.0782996170 OrangeFemale -0.021714183       1        L FALSE
#> 176...129   -0.0334858684 OrangeFemale -0.021714183       1        L FALSE
#> 198...130   -0.0866827714 OrangeFemale -0.021714183       1        L FALSE
#> 168.1...131 -0.0404910974 OrangeFemale -0.021714183       1        L FALSE
#> 174...132   -0.0697034848 OrangeFemale -0.021714183       1        L FALSE
#> 155.1...133 -0.0437682292 OrangeFemale -0.021714183       1        L FALSE
#> 194...134   -0.0696252457 OrangeFemale -0.021714183       1        L FALSE
#> 170...135   -0.0530803887 OrangeFemale -0.021714183       1        L FALSE
#> 187...136   -0.0667199838 OrangeFemale -0.021714183       1        L FALSE
#> 159.1...137 -0.0270699023 OrangeFemale -0.021714183       1        L FALSE
#> 173.1...138 -0.0587403636 OrangeFemale -0.021714183       1        L FALSE
#> 189...139   -0.0394899134 OrangeFemale -0.021714183       1        L FALSE
#> 154.2...140 -0.0440309646 OrangeFemale -0.021714183       1        L FALSE
#> 191.1...141 -0.0800380172 OrangeFemale -0.021714183       1        L FALSE
#> 166.1...142 -0.0462978876 OrangeFemale -0.021714183       1        L FALSE
#> 169.2...143 -0.0647722520 OrangeFemale -0.021714183       1        L FALSE
#> 166.2...144 -0.0462978876 OrangeFemale -0.021714183       1        L FALSE
#> 152...145   -0.0418830135 OrangeFemale -0.021714183       1        L FALSE
#> 176.1...146 -0.0334858684 OrangeFemale -0.021714183       1        L FALSE
#> 184.1...147 -0.0535585158 OrangeFemale -0.021714183       1        L FALSE
#> 196...148   -0.0849880686 OrangeFemale -0.021714183       1        L FALSE
#> 152.1...149 -0.0418830135 OrangeFemale -0.021714183       1        L FALSE
#> 200.1...150 -0.0759380502 OrangeFemale -0.021714183       1        L FALSE
#> 116...151   -0.0519359919   OrangeMale -0.021714183       1        L FALSE
#> 147...152   -0.0638050154   OrangeMale -0.021714183       1        L FALSE
#> 135...153   -0.0544807791   OrangeMale -0.021714183       1        L FALSE
#> 123...154   -0.0389899157   OrangeMale -0.021714183       1        L FALSE
#> 141...155   -0.0457032462   OrangeMale -0.021714183       1        L FALSE
#> 102...156   -0.0270359113   OrangeMale -0.021714183       1        L FALSE
#> 137...157   -0.0798972184   OrangeMale -0.021714183       1        L FALSE
#> 118...158   -0.0143616827   OrangeMale -0.021714183       1        L  TRUE
#> 142...159   -0.0504997403   OrangeMale -0.021714183       1        L FALSE
#> 127...160   -0.0332358914   OrangeMale -0.021714183       1        L FALSE
#> 114...161   -0.0443652846   OrangeMale -0.021714183       1        L FALSE
#> 141.1...162 -0.0457032462   OrangeMale -0.021714183       1        L FALSE
#> 105...163   -0.0653480748   OrangeMale -0.021714183       1        L FALSE
#> 142.1...164 -0.0504997403   OrangeMale -0.021714183       1        L FALSE
#> 114.1...165 -0.0443652846   OrangeMale -0.021714183       1        L FALSE
#> 138...166   -0.0426729958   OrangeMale -0.021714183       1        L FALSE
#> 149...167   -0.0603215604   OrangeMale -0.021714183       1        L FALSE
#> 105.1...168 -0.0653480748   OrangeMale -0.021714183       1        L FALSE
#> 143...169   -0.0529608225   OrangeMale -0.021714183       1        L FALSE
#> 141.2...170 -0.0457032462   OrangeMale -0.021714183       1        L FALSE
#> 120...171   -0.0539589545   OrangeMale -0.021714183       1        L FALSE
#> 117...172   -0.0404971139   OrangeMale -0.021714183       1        L FALSE
#> 111...173   -0.0239990063   OrangeMale -0.021714183       1        L FALSE
#> 129...174   -0.0694031121   OrangeMale -0.021714183       1        L FALSE
#> 145...175   -0.0345986882   OrangeMale -0.021714183       1        L FALSE
#> 127.1...176 -0.0332358914   OrangeMale -0.021714183       1        L FALSE
#> 130...177   -0.0602220450   OrangeMale -0.021714183       1        L FALSE
#> 134...178   -0.0580485771   OrangeMale -0.021714183       1        L FALSE
#> 127.2...179 -0.0332358914   OrangeMale -0.021714183       1        L FALSE
#> 126...180   -0.0379577927   OrangeMale -0.021714183       1        L FALSE
#> 101...181   -0.0379839038   OrangeMale -0.021714183       1        L FALSE
#> 103...182   -0.0336746052   OrangeMale -0.021714183       1        L FALSE
#> 147.1...183 -0.0638050154   OrangeMale -0.021714183       1        L FALSE
#> 139...184   -0.0580455688   OrangeMale -0.021714183       1        L FALSE
#> 111.1...185 -0.0239990063   OrangeMale -0.021714183       1        L FALSE
#> 133...186   -0.0169282967   OrangeMale -0.021714183       1        L  TRUE
#> 133.1...187 -0.0169282967   OrangeMale -0.021714183       1        L  TRUE
#> 120.1...188 -0.0539589545   OrangeMale -0.021714183       1        L FALSE
#> 141.3...189 -0.0457032462   OrangeMale -0.021714183       1        L FALSE
#> 128...190   -0.0517606514   OrangeMale -0.021714183       1        L FALSE
#> 145.1...191 -0.0345986882   OrangeMale -0.021714183       1        L FALSE
#> 128.1...192 -0.0517606514   OrangeMale -0.021714183       1        L FALSE
#> 146...193   -0.0762820766   OrangeMale -0.021714183       1        L FALSE
#> 130.1...194 -0.0602220450   OrangeMale -0.021714183       1        L FALSE
#> 122...195   -0.0477832589   OrangeMale -0.021714183       1        L FALSE
#> 148...196   -0.0715789940   OrangeMale -0.021714183       1        L FALSE
#> 136...197   -0.0678582767   OrangeMale -0.021714183       1        L FALSE
#> 121...198   -0.0341836714   OrangeMale -0.021714183       1        L FALSE
#> 101.1...199 -0.0379839038   OrangeMale -0.021714183       1        L FALSE
#> 129.1...200 -0.0694031121   OrangeMale -0.021714183       1        L FALSE
#> 186...201    0.0945703809 OrangeFemale  0.028269280       2        R  TRUE
#> 161...202    0.0351289019 OrangeFemale  0.028269280       2        R  TRUE
#> 191...203    0.0699859003 OrangeFemale  0.028269280       2        R  TRUE
#> 166...204    0.0806788980 OrangeFemale  0.028269280       2        R  TRUE
#> 185...205    0.1120916442 OrangeFemale  0.028269280       2        R  TRUE
#> 197...206    0.0896837235 OrangeFemale  0.028269280       2        R  TRUE
#> 156...207    0.0726631537 OrangeFemale  0.028269280       2        R  TRUE
#> 157...208    0.0854850270 OrangeFemale  0.028269280       2        R  TRUE
#> 184...209    0.0974977493 OrangeFemale  0.028269280       2        R  TRUE
#> 172...210    0.0718340279 OrangeFemale  0.028269280       2        R  TRUE
#> 192...211    0.1405441779 OrangeFemale  0.028269280       2        R  TRUE
#> 199...212    0.0726867641 OrangeFemale  0.028269280       2        R  TRUE
#> 154...213    0.0826977136 OrangeFemale  0.028269280       2        R  TRUE
#> 173...214    0.1080082167 OrangeFemale  0.028269280       2        R  TRUE
#> 186.1...215  0.0945703809 OrangeFemale  0.028269280       2        R  TRUE
#> 155...216    0.0535642416 OrangeFemale  0.028269280       2        R  TRUE
#> 200...217    0.1753446475 OrangeFemale  0.028269280       2        R  TRUE
#> 163...218    0.0993377446 OrangeFemale  0.028269280       2        R  TRUE
#> 153...219    0.0193730266 OrangeFemale  0.028269280       2        R FALSE
#> 185.1...220  0.1120916442 OrangeFemale  0.028269280       2        R  TRUE
#> 190...221    0.1244553738 OrangeFemale  0.028269280       2        R  TRUE
#> 168...222    0.1003142711 OrangeFemale  0.028269280       2        R  TRUE
#> 169...223    0.1059746352 OrangeFemale  0.028269280       2        R  TRUE
#> 165...224    0.0892869200 OrangeFemale  0.028269280       2        R  TRUE
#> 154.1...225  0.0826977136 OrangeFemale  0.028269280       2        R  TRUE
#> 169.1...226  0.1059746352 OrangeFemale  0.028269280       2        R  TRUE
#> 159...227    0.0929920199 OrangeFemale  0.028269280       2        R  TRUE
#> 177...228    0.1322491101 OrangeFemale  0.028269280       2        R  TRUE
#> 176...229    0.1133936796 OrangeFemale  0.028269280       2        R  TRUE
#> 198...230    0.0767979012 OrangeFemale  0.028269280       2        R  TRUE
#> 168.1...231  0.1003142711 OrangeFemale  0.028269280       2        R  TRUE
#> 174...232    0.0600230241 OrangeFemale  0.028269280       2        R  TRUE
#> 155.1...233  0.0535642416 OrangeFemale  0.028269280       2        R  TRUE
#> 194...234    0.0876822739 OrangeFemale  0.028269280       2        R  TRUE
#> 170...235    0.1004088531 OrangeFemale  0.028269280       2        R  TRUE
#> 187...236    0.1007409954 OrangeFemale  0.028269280       2        R  TRUE
#> 159.1...237  0.0929920199 OrangeFemale  0.028269280       2        R  TRUE
#> 173.1...238  0.1080082167 OrangeFemale  0.028269280       2        R  TRUE
#> 189...239    0.0890245735 OrangeFemale  0.028269280       2        R  TRUE
#> 154.2...240  0.0826977136 OrangeFemale  0.028269280       2        R  TRUE
#> 191.1...241  0.0699859003 OrangeFemale  0.028269280       2        R  TRUE
#> 166.1...242  0.0806788980 OrangeFemale  0.028269280       2        R  TRUE
#> 169.2...243  0.1059746352 OrangeFemale  0.028269280       2        R  TRUE
#> 166.2...244  0.0806788980 OrangeFemale  0.028269280       2        R  TRUE
#> 152...245    0.0181149655 OrangeFemale  0.028269280       2        R FALSE
#> 176.1...246  0.1133936796 OrangeFemale  0.028269280       2        R  TRUE
#> 184.1...247  0.0974977493 OrangeFemale  0.028269280       2        R  TRUE
#> 196...248    0.1439662443 OrangeFemale  0.028269280       2        R  TRUE
#> 152.1...249  0.0181149655 OrangeFemale  0.028269280       2        R FALSE
#> 200.1...250  0.1753446475 OrangeFemale  0.028269280       2        R  TRUE
#> 116...251   -0.0084621201   OrangeMale  0.028269280       2        L FALSE
#> 147...252   -0.0483983266   OrangeMale  0.028269280       2        L FALSE
#> 135...253   -0.0398974412   OrangeMale  0.028269280       2        L FALSE
#> 123...254   -0.0005468789   OrangeMale  0.028269280       2        L FALSE
#> 141...255   -0.0741304193   OrangeMale  0.028269280       2        L FALSE
#> 102...256   -0.0091279036   OrangeMale  0.028269280       2        L FALSE
#> 137...257   -0.0299873839   OrangeMale  0.028269280       2        L FALSE
#> 118...258   -0.0123922257   OrangeMale  0.028269280       2        L FALSE
#> 142...259   -0.0768649138   OrangeMale  0.028269280       2        L FALSE
#> 127...260   -0.0436590700   OrangeMale  0.028269280       2        L FALSE
#> 114...261   -0.0239782237   OrangeMale  0.028269280       2        L FALSE
#> 141.1...262 -0.0741304193   OrangeMale  0.028269280       2        L FALSE
#> 105...263    0.0034830176   OrangeMale  0.028269280       2        L FALSE
#> 142.1...264 -0.0768649138   OrangeMale  0.028269280       2        L FALSE
#> 114.1...265 -0.0239782237   OrangeMale  0.028269280       2        L FALSE
#> 138...266   -0.0436025411   OrangeMale  0.028269280       2        L FALSE
#> 149...267   -0.0247822402   OrangeMale  0.028269280       2        L FALSE
#> 105.1...268  0.0034830176   OrangeMale  0.028269280       2        L FALSE
#> 143...269   -0.0452659272   OrangeMale  0.028269280       2        L FALSE
#> 141.2...270 -0.0741304193   OrangeMale  0.028269280       2        L FALSE
#> 120...271   -0.0140659553   OrangeMale  0.028269280       2        L FALSE
#> 117...272   -0.0020240245   OrangeMale  0.028269280       2        L FALSE
#> 111...273    0.0084446797   OrangeMale  0.028269280       2        L FALSE
#> 129...274   -0.0211886992   OrangeMale  0.028269280       2        L FALSE
#> 145...275   -0.0745900619   OrangeMale  0.028269280       2        L FALSE
#> 127.1...276 -0.0436590700   OrangeMale  0.028269280       2        L FALSE
#> 130...277   -0.0562999118   OrangeMale  0.028269280       2        L FALSE
#> 134...278   -0.0410066027   OrangeMale  0.028269280       2        L FALSE
#> 127.2...279 -0.0436590700   OrangeMale  0.028269280       2        L FALSE
#> 126...280   -0.0331382592   OrangeMale  0.028269280       2        L FALSE
#> 101...281   -0.0158937191   OrangeMale  0.028269280       2        L FALSE
#> 103...282    0.0030086091   OrangeMale  0.028269280       2        L FALSE
#> 147.1...283 -0.0483983266   OrangeMale  0.028269280       2        L FALSE
#> 139...284   -0.0261393056   OrangeMale  0.028269280       2        L FALSE
#> 111.1...285  0.0084446797   OrangeMale  0.028269280       2        L FALSE
#> 133...286   -0.0424471942   OrangeMale  0.028269280       2        L FALSE
#> 133.1...287 -0.0424471942   OrangeMale  0.028269280       2        L FALSE
#> 120.1...288 -0.0140659553   OrangeMale  0.028269280       2        L FALSE
#> 141.3...289 -0.0741304193   OrangeMale  0.028269280       2        L FALSE
#> 128...290   -0.0572383853   OrangeMale  0.028269280       2        L FALSE
#> 145.1...291 -0.0745900619   OrangeMale  0.028269280       2        L FALSE
#> 128.1...292 -0.0572383853   OrangeMale  0.028269280       2        L FALSE
#> 146...293   -0.0629079831   OrangeMale  0.028269280       2        L FALSE
#> 130.1...294 -0.0562999118   OrangeMale  0.028269280       2        L FALSE
#> 122...295   -0.0287924851   OrangeMale  0.028269280       2        L FALSE
#> 148...296   -0.0310355942   OrangeMale  0.028269280       2        L FALSE
#> 136...297   -0.0214255473   OrangeMale  0.028269280       2        L FALSE
#> 121...298   -0.0211528573   OrangeMale  0.028269280       2        L FALSE
#> 101.1...299 -0.0158937191   OrangeMale  0.028269280       2        L FALSE
#> 129.1...300 -0.0211886992   OrangeMale  0.028269280       2        L FALSE
#> 99...301    -0.0990600036   BlueFemale -0.001016054       3        L FALSE
#> 80...302    -0.0299141396   BlueFemale -0.001016054       3        L FALSE
#> 89...303    -0.0316605756   BlueFemale -0.001016054       3        L FALSE
#> 69...304    -0.0462239246   BlueFemale -0.001016054       3        L FALSE
#> 89.1...305  -0.0316605756   BlueFemale -0.001016054       3        L FALSE
#> 77...306    -0.0240326620   BlueFemale -0.001016054       3        L FALSE
#> 96...307    -0.0394080966   BlueFemale -0.001016054       3        L FALSE
#> 60...308    -0.0115237407   BlueFemale -0.001016054       3        L FALSE
#> 65...309    -0.0910552724   BlueFemale -0.001016054       3        L FALSE
#> 55...310     0.0035906133   BlueFemale -0.001016054       3        L  TRUE
#> 61...311    -0.0211732472   BlueFemale -0.001016054       3        L FALSE
#> 99.1...312  -0.0990600036   BlueFemale -0.001016054       3        L FALSE
#> 65.1...313  -0.0910552724   BlueFemale -0.001016054       3        L FALSE
#> 87...314    -0.0580327475   BlueFemale -0.001016054       3        L FALSE
#> 90...315    -0.0425563435   BlueFemale -0.001016054       3        L FALSE
#> 56...316    -0.0225920819   BlueFemale -0.001016054       3        L FALSE
#> 62...317    -0.0332043815   BlueFemale -0.001016054       3        L FALSE
#> 64...318    -0.0539135167   BlueFemale -0.001016054       3        L FALSE
#> 56.1...319  -0.0225920819   BlueFemale -0.001016054       3        L FALSE
#> 56.2...320  -0.0225920819   BlueFemale -0.001016054       3        L FALSE
#> 100...321   -0.0474439279   BlueFemale -0.001016054       3        L FALSE
#> 100.1...322 -0.0474439279   BlueFemale -0.001016054       3        L FALSE
#> 57...323    -0.0302150328   BlueFemale -0.001016054       3        L FALSE
#> 96.1...324  -0.0394080966   BlueFemale -0.001016054       3        L FALSE
#> 79...325    -0.0620159893   BlueFemale -0.001016054       3        L FALSE
#> 70...326    -0.0900308008   BlueFemale -0.001016054       3        L FALSE
#> 73...327    -0.0540653166   BlueFemale -0.001016054       3        L FALSE
#> 86...328    -0.0671394398   BlueFemale -0.001016054       3        L FALSE
#> 55.1...329   0.0035906133   BlueFemale -0.001016054       3        L  TRUE
#> 67...330    -0.0485218873   BlueFemale -0.001016054       3        L FALSE
#> 85...331    -0.0526105396   BlueFemale -0.001016054       3        L FALSE
#> 66...332    -0.0267729912   BlueFemale -0.001016054       3        L FALSE
#> 92...333    -0.0761981287   BlueFemale -0.001016054       3        L FALSE
#> 61.1...334  -0.0211732472   BlueFemale -0.001016054       3        L FALSE
#> 75...335    -0.0213147205   BlueFemale -0.001016054       3        L FALSE
#> 64.1...336  -0.0539135167   BlueFemale -0.001016054       3        L FALSE
#> 60.1...337  -0.0115237407   BlueFemale -0.001016054       3        L FALSE
#> 98...338    -0.1152517953   BlueFemale -0.001016054       3        L FALSE
#> 67.1...339  -0.0485218873   BlueFemale -0.001016054       3        L FALSE
#> 74...340    -0.0237612549   BlueFemale -0.001016054       3        L FALSE
#> 52...341    -0.0176837238   BlueFemale -0.001016054       3        L FALSE
#> 78...342    -0.0575669010   BlueFemale -0.001016054       3        L FALSE
#> 83...343    -0.0713853763   BlueFemale -0.001016054       3        L FALSE
#> 80.1...344  -0.0299141396   BlueFemale -0.001016054       3        L FALSE
#> 67.2...345  -0.0485218873   BlueFemale -0.001016054       3        L FALSE
#> 95...346    -0.0869163604   BlueFemale -0.001016054       3        L FALSE
#> 82...347    -0.0737636554   BlueFemale -0.001016054       3        L FALSE
#> 66.1...348  -0.0267729912   BlueFemale -0.001016054       3        L FALSE
#> 70.1...349  -0.0900308008   BlueFemale -0.001016054       3        L FALSE
#> 59...350    -0.0334997903   BlueFemale -0.001016054       3        L FALSE
#> 44...351     0.0162715373     BlueMale -0.001016054       3        R  TRUE
#> 48...352     0.1174257811     BlueMale -0.001016054       3        R  TRUE
#> 29...353     0.0306481161     BlueMale -0.001016054       3        R  TRUE
#> 17...354     0.0388325191     BlueMale -0.001016054       3        R  TRUE
#> 50...355     0.1123263324     BlueMale -0.001016054       3        R  TRUE
#> 12...356    -0.0141510393     BlueMale -0.001016054       3        R FALSE
#> 31...357     0.0344844002     BlueMale -0.001016054       3        R  TRUE
#> 6...358      0.0204005730     BlueMale -0.001016054       3        R  TRUE
#> 25...359     0.0446254895     BlueMale -0.001016054       3        R  TRUE
#> 5...360      0.0261078632     BlueMale -0.001016054       3        R  TRUE
#> 9...361      0.0151422260     BlueMale -0.001016054       3        R  TRUE
#> 15...362     0.0063533301     BlueMale -0.001016054       3        R  TRUE
#> 30...363     0.0823477363     BlueMale -0.001016054       3        R  TRUE
#> 37...364     0.0649354196     BlueMale -0.001016054       3        R  TRUE
#> 9.1...365    0.0151422260     BlueMale -0.001016054       3        R  TRUE
#> 44.1...366   0.0162715373     BlueMale -0.001016054       3        R  TRUE
#> 36...367     0.0671940914     BlueMale -0.001016054       3        R  TRUE
#> 39...368     0.0783033376     BlueMale -0.001016054       3        R  TRUE
#> 8...369      0.0428495675     BlueMale -0.001016054       3        R  TRUE
#> 18...370     0.0240048615     BlueMale -0.001016054       3        R  TRUE
#> 34...371     0.0442568627     BlueMale -0.001016054       3        R  TRUE
#> 27...372     0.0603716866     BlueMale -0.001016054       3        R  TRUE
#> 18.1...373   0.0240048615     BlueMale -0.001016054       3        R  TRUE
#> 13...374     0.0568630038     BlueMale -0.001016054       3        R  TRUE
#> 3...375      0.0124171860     BlueMale -0.001016054       3        R  TRUE
#> 12.1...376  -0.0141510393     BlueMale -0.001016054       3        R FALSE
#> 45...377     0.1254157448     BlueMale -0.001016054       3        R  TRUE
#> 41...378     0.0709894706     BlueMale -0.001016054       3        R  TRUE
#> 38...379     0.0804937541     BlueMale -0.001016054       3        R  TRUE
#> 8.1...380    0.0428495675     BlueMale -0.001016054       3        R  TRUE
#> 7...381     -0.0095654404     BlueMale -0.001016054       3        R FALSE
#> 49...382     0.0973413905     BlueMale -0.001016054       3        R  TRUE
#> 22...383     0.0625347528     BlueMale -0.001016054       3        R  TRUE
#> 24...384     0.0646569140     BlueMale -0.001016054       3        R  TRUE
#> 9.2...385    0.0151422260     BlueMale -0.001016054       3        R  TRUE
#> 30.1...386   0.0823477363     BlueMale -0.001016054       3        R  TRUE
#> 14...387     0.0392251476     BlueMale -0.001016054       3        R  TRUE
#> 12.2...388  -0.0141510393     BlueMale -0.001016054       3        R FALSE
#> 35...389     0.0708289583     BlueMale -0.001016054       3        R  TRUE
#> 15.1...390   0.0063533301     BlueMale -0.001016054       3        R  TRUE
#> 41.1...391   0.0709894706     BlueMale -0.001016054       3        R  TRUE
#> 5.1...392    0.0261078632     BlueMale -0.001016054       3        R  TRUE
#> 42...393     0.0607561243     BlueMale -0.001016054       3        R  TRUE
#> 22.1...394   0.0625347528     BlueMale -0.001016054       3        R  TRUE
#> 38.1...395   0.0804937541     BlueMale -0.001016054       3        R  TRUE
#> 34.1...396   0.0442568627     BlueMale -0.001016054       3        R  TRUE
#> 23...397     0.0761194978     BlueMale -0.001016054       3        R  TRUE
#> 32...398     0.0336461772     BlueMale -0.001016054       3        R  TRUE
#> 1...399      0.0131494769     BlueMale -0.001016054       3        R  TRUE
#> 11...400     0.0101110325     BlueMale -0.001016054       3        R  TRUE