Parkinson data set

Parkinson data set

Usage

data(parkinson)

Format

A data set containing 195 observations from 2 parkinson types.

Type

Class variable has 2 classes, there are 48 cases of healthy people and 147 cases with Parkinson. The feature variables are biomedical voice measures.

X1

Average vocal fundamental frequency.

X2

Maximum vocal fundamental frequency.

X3

Minimum vocal fundamental frequency.

X4

MDVP:Jitter(%) measures of variation in fundamental frequency.

X5

MDVP:Jitter(Abs) measures of variation in fundamental frequency.

X6

MDVP:RAP measures of variation in fundamental frequency.

X7

MDVP:PPQ measures of variation in fundamental frequency.

X8

Jitter:DDP measures of variation in fundamental frequency.

X9

MDVP:Shimmer measures of variation in amplitude.

X10

MDVP:Shimmer(dB) measures of variation in amplitude.

X11

Shimmer:APQ3 measures of variation in amplitude.

X12

Shimmer:APQ5 measures of variation in amplitude.

X13

MDVP:APQ measures of variation in amplitude.

X14

Shimmer:DDA measures of variation in amplitude.

X15

NHR measures of ratio of noise to tonal components in the voice.

X16

HNR measures of ratio of noise to tonal components in the voice.

X17

RPDE nonlinear dynamical complexity measures.

X18

D2 nonlinear dynamical complexity measures.

X19

DFA - Signal fractal scaling exponent.

X20

spread1 Nonlinear measures of fundamental frequency variation.

X21

spread2 Nonlinear measures of fundamental frequency variation.

X22

PPE Nonlinear measures of fundamental frequency variation.

A data frame with 195 rows and 23 variables

Source

<https://archive.ics.uci.edu/ml/datasets/Parkinsons>